% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sim_results.R
\name{analyse_sim_results}
\alias{analyse_sim_results}
\title{Analyse superNOVA simulated data results for two group and three group simulated data.}
\usage{
analyse_sim_results(ddcor_res, true_pairs, sig_cutoff = 0.05,
  verbose = F)
}
\arguments{
\item{ddcor_res}{The flattened chow_result file}

\item{true_pairs}{The true_pairs table from simulate_data() or simulate_data_3gr().}

\item{sig_cutoff}{The significance p-value (alpha) to use to call observed results as significant.}

\item{verbose}{Whether to print additional output or not.}
}
\value{
Returns a list, including an expected versus observed table for the simulated results, and stats including TPR, TNR, FPR, FNR.
}
\description{
Provides a way to merge expected and observed differential coexpressed tables and calculate the sensitivity/specificity/FPR/FNR.
}
\keyword{analyse}
\keyword{simulate}
\keyword{superNOVA}
