% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{b0}
\alias{b0}
\alias{b0.default}
\alias{b0.formula}
\title{b0}
\usage{
b0(fit, ...)

\method{b0}{default}(fit, ...)

\method{b0}{formula}(formula, data = list(), ...)
}
\arguments{
\item{fit}{A \code{\link{lm}} object.}

\item{...}{Passthrough arguments.}

\item{formula}{A \code{\link{formula}}.}

\item{data}{A \code{\link{data.frame}}.}
}
\value{
The intercept of the relevant model.
}
\description{
A function to extract intercept/beta0 value.
}
