% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{b1}
\alias{b1}
\alias{b1.default}
\alias{b1.formula}
\title{b1}
\usage{
b1(fit, ...)

\method{b1}{default}(fit, ...)

\method{b1}{formula}(formula, data = list(), ...)
}
\arguments{
\item{fit}{A \code{\link{lm}} object.}

\item{...}{Passthrough arguments.}

\item{formula}{A \code{\link{formula}}.}

\item{data}{A \code{\link{data.frame}}.}
}
\value{
The slope of the relevant model.
}
\description{
A function to extract slope/beta1 value.
}
