% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{barHelp}
\alias{barHelp}
\title{barHelp}
\usage{
barHelp(x, y)
}
\arguments{
\item{x}{Original name for row.}

\item{y}{Number of spaces to add.}
}
\value{
The original row name with a number of spaces and vertical bar added.
}
\description{
A function to produce vertical bar separators.
}
