% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_basis_freq}
\alias{calc_basis_freq}
\title{Calculate the frequency basis functions are used in the folds}
\usage{
calc_basis_freq(fold_basis_results, n_folds)
}
\arguments{
\item{fold_basis_results}{List of lists of variable sets found in the basis
function data-adaptive parameter section}

\item{n_folds}{Number of folds used}
}
\description{
Gives information on which variable sets are used in the
basis function in the data-adaptive estimation section
}
