% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_CIs}
\alias{calc_CIs}
\title{Calculate Confidence Interals}
\usage{
calc_CIs(psi, psi_se)
}
\arguments{
\item{psi}{The Psi parameter calculated}

\item{psi_se}{Standard deviation for the Psi parameter}
}
\description{
Gives information on which variable sets are used in the
basis function in the data-adaptive estimation section
}
