% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_utils.R
\name{calc_final_effect_mod_param}
\alias{calc_final_effect_mod_param}
\title{Calculates the Effect Modification Shift Parameter}
\usage{
calc_final_effect_mod_param(
  tmle_fit_av,
  tmle_fit_at,
  exposure,
  at,
  av,
  effect_m_name,
  fold_k,
  em_learner
)
}
\arguments{
\item{tmle_fit_av}{TMLE results for the validation fold}

\item{tmle_fit_at}{TMLE results for the training fold}

\item{exposure}{The exposure identified}

\item{at}{Training data}

\item{av}{Validation data}

\item{effect_m_name}{The effect modifier}

\item{fold_k}{The fold the effect modification was found}

\item{em_learner}{The Super Learner of decision trees}
}
\description{
Use a decision tree estimator to regress the eif of an exposure
on to a covariate - both of these have been data adaptively determined.
}
