% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_utils.R
\name{calc_final_ind_shift_param}
\alias{calc_final_ind_shift_param}
\title{Calculate Final Results}
\usage{
calc_final_ind_shift_param(indiv_shift_fold_results, exposure, fold_k)
}
\description{
Calculate Final Results
}
\details{
Set of functions that calculate the final target parameter
depending on single variable shift, effect modification shift,
or a joint stochastic intervention.
}
