% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_intxn_results}
\alias{calc_intxn_results}
\title{Calculates the Interaction Parameter}
\usage{
calc_intxn_results(results_table, joint_shift_fold_results)
}
\description{
Using the output results for the eif use the delta method
to simply subtract the two additive estimates from the joint shift. Do
the same for the eif to get the variance for this new parameter.
}
