% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_utils.R
\name{calc_mediation_param}
\alias{calc_mediation_param}
\title{Calculates the Mediation Shift Parameters}
\usage{
calc_mediation_param(
  tmle_fit_a_shift,
  tmle_fit_a_z_shift,
  exposure,
  mediator,
  fold_k,
  y
)
}
\description{
Estimates the shift parameter for a natural direct and indirect
effect
}
