% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pooled_em_shifts.R
\name{calc_pooled_em_shifts}
\alias{calc_pooled_em_shifts}
\title{Compute the Pooled Shift Parameter Estimate From the Fold Specific Results
For the Effect Modification Parameter}
\usage{
calc_pooled_em_shifts(
  y,
  em_shift_results,
  estimator = c("tmle", "onestep"),
  fluc_mod_out = NULL,
  w_names,
  a_names,
  fluctuation
)
}
\arguments{
\item{y}{A \code{numeric} vector of the observed outcomes.}

\item{em_shift_results}{Shift results for effect modification found across
the folds.}

\item{estimator}{Type of estimator to use (TMLE or one-step)}

\item{fluc_mod_out}{An object giving values of the logistic tilting model
for targeted minimum loss estimation. This type of object should be the
output of the internal routines to perform this step of the TML estimation
procedure, as given by \code{\link{fit_fluctuation}}.}

\item{w_names}{Names of covariates}

\item{a_names}{Names of exposure}

\item{fluctuation}{type of fluctuation to use}
}
\value{
A \code{list} containing the parameter estimate, estimated variance
based on the efficient influence function (EIF), the estimate of the EIF
incorporating inverse probability of censoring weights, and the estimate of
the EIF without the application of such weights.
}
\description{
Compute the Pooled Shift Parameter Estimate From the Fold Specific Results
For the Effect Modification Parameter
}
\details{
Estimate the value of the pooled causal parameter alongside
statistical inference for the parameter estimate based on the nuisance
parameters from the fold specific results.
}
