% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pooled_med_shifts.R
\name{calc_pooled_med_shifts}
\alias{calc_pooled_med_shifts}
\title{Compute the Pooled Mediation Shift Parameter Estimate From the Fold
Specific Results}
\usage{
calc_pooled_med_shifts(
  med_shift_results,
  estimator = c("onestep"),
  fluc_mod_out = NULL,
  a_names,
  w_names,
  z_names,
  y_name,
  fluctuation
)
}
\arguments{
\item{med_shift_results}{List of data frames of the mediation results across
the parallelized CV procedure}

\item{estimator}{The type of estimator to be fit, either \code{"tmle"} for
targeted maximum likelihood estimation or \code{"onestep"} for a one-step
estimator.}

\item{fluc_mod_out}{An object giving values of the logistic tilting model
for targeted minimum loss estimation. This type of object should be the
output of the internal routines to perform this step of the TML estimation
procedure, as given by \code{\link{fit_fluctuation}}.}

\item{a_names}{List of exposure names}

\item{w_names}{List of covariate names}

\item{z_names}{List of mediator names}

\item{y_name}{Outcome name}

\item{fluctuation}{Type of fluctuation to be used}
}
\value{
A \code{list} containing the parameter estimate, estimated variance
based on the efficient influence function (EIF), the estimate of the EIF
incorporating inverse probability of censoring weights, and the estimate of
the EIF without the application of such weights.
}
\description{
Compute the Pooled Mediation Shift Parameter Estimate From the Fold
Specific Results
}
\details{
Estimate the value of the pooled causal parameter alongside statistical
inference for the parameter estimate based on the nuisance parameters from
the fold specific results for the mediation parameter.
}
