% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{calc_pre}
\alias{calc_pre}
\title{Calculate the PRE for a model}
\usage{
calc_pre(fit_augmented, fit_compact)
}
\arguments{
\item{fit_augmented}{An \code{\link{lm}} object of the model with the predictors being tested}

\item{fit_compact}{An \code{\link{lm}} object of the model to test against}
}
\value{
The PRE for the model.
}
\description{
Calculate the PRE for a model
}
