% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_pvals}
\alias{calc_pvals}
\title{Calculate the p-values based on variance of the influence curve}
\usage{
calc_pvals(psi, variance)
}
\arguments{
\item{psi}{The estimated target parameter}

\item{variance}{The variance of that parameter}
}
\description{
Calculates the p-value using the standard formula
}
