% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{calc_ssr}
\alias{calc_ssr}
\title{Calculate the SSR for a model}
\usage{
calc_ssr(fit)
}
\arguments{
\item{fit}{An \code{\link{lm}} object.}
}
\value{
The sum of squares regression/between/explained
}
\description{
Calculate the SSR for a model
}
