% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_line}
\alias{cat_line}
\title{Paste, Concatenate, add EOL and Print}
\usage{
cat_line(...)
}
\arguments{
\item{...}{Character vectors to paste together.}
}
\value{
None (invisible \code{NULL}).
}
\description{
Paste, Concatenate, add EOL and Print
}
\keyword{internal}
