% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chowCor.R
\name{chowCor}
\alias{chowCor}
\title{Runs the superNOVA pipeline directly on one or two feature matrices.}
\usage{
chowCor(design_mat, matA, matB = NULL, compare = NULL,
  corrType = "pearson")
}
\arguments{
\item{design_mat}{A model.matrix type design matrix, denoting which samples are in which subgroups of the dataset. Required.}

\item{matA}{A features (rows) by samples (columns) data.frame of feature values, such as a gene expression matrix. Required.}

\item{matB}{A second features (rows) by samples (columns) data.frame of feature values, such as a gene expression matrix. Values in this matrix will be compared to matA if provided. Default=NULL.}

\item{compare}{Which groups of samples in the design matrix should we evaluate? Otherwise, all groups in the design matrix will be evaluated. Default=NULL.}

\item{corrType}{The base correlation metric to evaluate coexpression. One of "pearson","spearman", or "bicor". Default=pearson.}
}
\value{
Returns a list of matrices, includes pvalues for the superNOVA test (pvalues) and their classes (classes), correlation values for each subgroup model (corrs) and their pvalues (corrsP), correlation values for the global model (globalCor) and its pvalues (globalCorP), and a flag to indicate whether a second matrix was used (secondMat).
}
\description{
Evaluates differential coexpression between two or more subgroups of samples in the data versus the global model.
}
\keyword{superNOVA}
