% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chowParallel.R
\name{chowParallel}
\alias{chowParallel}
\title{Calls the superNOVA pipeline, splitting input matrix into multiple batch jobs on an HPC.}
\usage{
chowParallel(inputMat, design, outputFile, compare = NULL,
  sigOutput = FALSE, sigThresh = 0.05, verbose = FALSE,
  corrType = "pearson", perBatch = 10, coresPerJob = 2,
  timePerJob = 60, memPerJob = 2000,
  batchConfig = system.file("config/batchConfig_Zhang.R", package =
  "superNOVA"), batchDir = "batchRegistry", batchWarningLevel = 0,
  batchSeed = 12345, maxRetries = 3, testJob = FALSE,
  chunkSize = 1)
}
\arguments{
\item{inputMat}{The matrix (or data.frame) of values (e.g., gene expression values from an RNA-seq or microarray study) that you are interested in analyzing. The rownames of this matrix should correspond to the identifiers whose correlations and differential correlations you are interested in analyzing, while the columns should correspond to the rows of the design matrix and should be separable into your compare.}

\item{design}{A standard model.matrix created design matrix. Rows correspond to samples and colnames refer to the names of the conditions that you are interested in analyzing. Only 0's or 1's are allowed in the design matrix. Please see vignettes for more information.}

\item{outputFile}{Location to save the output. Required.}

\item{compare}{Vector of two character strings, each corresponding to one group name in the design matrix, that should be compared.}

\item{sigOutput}{Should we save the significant results in a separate file? Default = FALSE.}

\item{sigThresh}{This numeric value specifies the p-value threshold at which a differential correlation p-value is deemed significant for differential correlation class calculation. Default = 1, as investigators may use different cutoff thresholds; however, this can be lowered to establish significant classes as desired.}

\item{verbose}{Option indicating whether the program should give more frequent updates about its operations. Default = FALSE.}

\item{corrType}{The correlation type of the analysis, limited to "pearson","spearman",or "bicor". Default = "pearson".}

\item{perBatch}{Number of times to split the features of the input data into separate batches. A higher number creates a larger number of jobs, but may be less uniform. Default = 10.}

\item{coresPerJob}{Number of cores to use on each batch job run. Default = 2.}

\item{timePerJob}{Walltime to request for each batch job (e.g. in a HPC cluster), in minutes. Default = 60}

\item{memPerJob}{Memory to request for each batch job (e.g. in a HPC cluster), in MB. Default = 2000}

\item{batchConfig}{Location of the batchtools configuration file (e.g. to configure this tool to work with your HPC cluster). Defaults to one used at inst/config/batchConfig_Zhang.R.}

\item{batchDir}{Location to store temporary files, logs, and results of the batch run. This is the registry for the batchtools R package. Default = batchRegistry/}

\item{batchWarningLevel}{Warning level on remote nodes during chowCor calculation (equivalent to setting options(warn=batchWarningLevel). Default = 0.}

\item{batchSeed}{Random seed to use on all batch jobs. Default = 12345.}

\item{maxRetries}{Number of times to re-submit jobs that failed. This is helpful for jobs that failed due to transient errors on an HPC. Default = 3}

\item{testJob}{Test one job before running it? Default = FALSE}

\item{chunkSize}{Execute multiple splits sequentially on each node. Default = 1 (false)}
}
\value{
Returns whether all jobs successfully executed or not. Output is in the output file.
}
\description{
Evaluates differential coexpression between two or more subgroups of samples in the data versus the global model, using multiple nodes in parallel in a batch environment.
}
\keyword{superNOVA}
