% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chowParallelWorker.R
\name{chowParallelWorker}
\alias{chowParallelWorker}
\title{Worker thread for chowParallel (internal function).}
\usage{
chowParallelWorker(job, data, instance)
}
\arguments{
\item{job}{A batchtools job instance.}

\item{data}{A named list containing the program kwargs.}

\item{instance}{Required by batchtools, but not used currently.}
}
\value{
Results from running chowCor() on data block A and B. See chowCor() documentation for more information.
}
\description{
Runs chowCor between two or more subgroups of samples in the data versus the global model on a worker thread for two blocks of data (A and B).
}
\keyword{superNOVA}
