% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_data}
\alias{class_data}
\title{Teacher effectiveness data}
\format{
A data frame with 105 observations on the following 2 variables:
\describe{
\item{\code{teacher}}{coded as A, B, or C}
\item{\code{outcome}}{each student's score on the outcome test}
}
}
\usage{
class_data
}
\description{
These are hypothetical data for a small study comparing the effectiveness of three different
teachers. Each teacher was randomly assigned a group of 35 high school students and asked to
teach a 5-day unit on a new science topic for which none of the students had any prior knowledge.
All three teachers used the same curriculum materials and lesson plans, and student learning was
assessed using a common 30-item test at the end of the unit. The research question was: Were some
teachers more effective than others? Did teachers differ in the amount of learning they were able
to produce in their students?
}
\keyword{datasets}
