% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_folds.R
\name{create_cv_folds}
\alias{create_cv_folds}
\title{Stratified CV to insure balance (by one grouping variable, Y)}
\usage{
create_cv_folds(V, Y, verbose = F)
}
\arguments{
\item{verbose}{If T will display extra output.}

\item{v}{number of folds}

\item{y}{Outcome variable. If binary will be used for stratification.}
}
\value{
Vector of fold assignments.
}
\description{
Creates a dummy variable that partitions the data into v
equal sized groups for v-fold CV.
}
