% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_sls.R
\name{create_sls}
\alias{create_sls}
\title{Create default Super Learner estimators for the data adaptive
and nuisance parameters used in \code{SuperNOVA}}
\usage{
create_sls()
}
\value{
List of ensemble estimators
}
\description{
If Super Learners are not passed to the stack arguments
this function is used to create some default ensemble machine
learning estimators for each parameter. The default estimators are
fast but also flexible.
}
