% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_models.R
\name{drop_term}
\alias{drop_term}
\title{Drop a term from the given model}
\usage{
drop_term(fit, term)
}
\arguments{
\item{fit}{The model to update.}

\item{term}{The term to drop from the model.}
}
\value{
An object of the class \code{lm}.
}
\description{
This function is needed to re-fit the models for Type III SS. If you have a model with an
interactive term (e.g. \code{y ~ a + b + a:b}), when you try to refit without one of the lower-order
terms (e.g. \code{y ~ a +    a:b}) \code{\link[=lm]{lm()}} will add it back in. This function uses a fitting function
that operates underneath \code{lm()} to circumvent this behavior. (It is very similar to \code{\link[=drop1]{drop1()}}.)
}
