% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eifs.R
\name{eif}
\alias{eif}
\title{Compute the Shift Parameter Estimate and the Efficient Influence Function}
\usage{
eif(
  y,
  qn,
  qn_unscaled,
  hn,
  estimator = c("tmle", "onestep"),
  fluc_mod_out = NULL
)
}
\arguments{
\item{y}{A \code{numeric} vector of the observed outcomes.}

\item{qn}{An object providing the value of the outcome evaluated after
imposing a shift in the treatment. This object is passed in after being
constructed by a call to the internal function \code{est_Q}. These values
have been scaled to 0-1.}

\item{qn_unscaled}{Same as qn but not scaled - kept in original outcome
scale.}

\item{hn}{An object providing values of the auxiliary ("clever") covariate,
constructed from the treatment mechanism and required for targeted minimum
loss-based estimation. This object object should be passed in after being
constructed by a call to the internal function \code{est_Hn}.}

\item{estimator}{The type of estimator to be fit, either \code{"tmle"} for
targeted maximum likelihood estimation or \code{"onestep"} for a one-step
estimator.}

\item{fluc_mod_out}{An object giving values of the logistic tilting model
for targeted minimum loss estimation. This type of object should be the
output of the internal routines to perform this step of the TML estimation
procedure, as given by \code{\link{fit_fluctuation}}.}
}
\value{
A \code{list} containing the parameter estimate, estimated variance
based on the efficient influence function (EIF), the estimate of the EIF
incorporating inverse probability of censoring weights, and the estimate of
the EIF without the application of such weights.
}
\description{
Compute the Shift Parameter Estimate and the Efficient Influence Function
}
\details{
Estimate the value of the causal parameter alongside statistical
inference for the parameter estimate based on the efficient influence
function of the target parameter.
}
