% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R
\name{equation}
\alias{equation}
\title{Print the output of \code{lm()} with the fitted equation.}
\usage{
equation(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{The fitted linear model to print.}

\item{digits}{The minimal number of significant digits.}
}
\value{
Invisibly return the fitted linear model.
}
\description{
Print the output of \code{lm()} with the fitted equation.
}
