% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_nde.R
\name{est_nde}
\alias{est_nde}
\title{Estimate the NDE nuisance parameters for each mixture
interaction identified}
\usage{
est_nde(
  at,
  av,
  w,
  a,
  z,
  y,
  q_ests,
  e_ests,
  g_ests,
  psi_learner,
  family,
  parallel_cv,
  seed,
  h_aw_trunc_lvl = h_aw_trunc_lvl
)
}
\arguments{
\item{at}{Training data}

\item{av}{Validation data}

\item{w}{Vector of characters denoting covariates}

\item{z}{Vector of characters denoting mediators}

\item{y}{The outcome variable}

\item{family}{Binomial or continuous}

\item{parallel_cv}{TRUE/FALSE if cv parallelization is used}

\item{seed}{Seed number}

\item{h_aw_trunc_lvl}{Truncation level of the clever covariate (induces more
bias to reduce variance)}

\item{no_mix_rules}{TRUE/FALSE indicator for if no mixture rules were found}

\item{aw_stack}{Super Learner library for fitting Q (outcome mechanism) and
g (treatment mechanism)}

\item{psi_z_stack}{Super Learner library for fitting Q_diff (difference
in conditional outcomes for A=1 and A = 0, and Z = z, W = w) onto W of
controls observations}

\item{rules}{Dataframe of rules found during the PRE fitting process}
}
\value{
A list of dataframes where the nuisance parameters are added to
the raw data.
}
\description{
For each mixture mixture interaction found, create a g estimator
for the probability of being exposed to the rule thresholds,
a Q estimator for the outcome E(Y| A = a_mix, Z, W), an e estimator for
e(A|Z,W) and a Phi estimator that regresses the outcome counterfactuals on
the covariates of A =0.  Get estimates of g, e,phi, and Q using the
validation data and calculate the clever covariate used in the
TMLE fluctuation step.
}
