% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_Q_w_shifted_mediation.R
\name{est_Q_w_shifted_mediation}
\alias{est_Q_w_shifted_mediation}
\title{Estimate the Outcome Mechanism with Shifted A and Z}
\usage{
est_Q_w_shifted_mediation(
  exposure,
  mediator,
  delta,
  mu_learner,
  covars,
  av,
  at,
  upper_bound = upper_bound,
  lower_bound = lower_bound
)
}
\arguments{
\item{exposure}{A \code{character} vector of exposures to be shifted.}

\item{mediator}{The mediator variable}

\item{delta}{A \code{numeric} indicating the magnitude of the shift to be
computed for the exposure \code{A}. This is passed to the internal
\code{\link{shift_additive}} and is currently limited to additive shifts.}

\item{mu_learner}{Object containing a set of instantiated learners from the
\pkg{sl3}, to be used in fitting an ensemble model.}

\item{covars}{A \code{character} vector covariates to adjust for.}

\item{av}{A \code{dataframe} of validation data specific to the fold}

\item{at}{A \code{dataframe} of training data specific to the fold}

\item{zn_estim}{Density estimates of the mediator under various shifts}
}
\value{
A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W).
}
\description{
Estimate the Outcome Mechanism with Shifted A and Z
}
\details{
Compute the outcome regression for the observed data, including
with the shift imposed by the intervention. This returns the outcome
regression for the observed data (at A) and under the counterfactual shift
shift (at A + delta).
}
