% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{find_categorical_vars}
\alias{find_categorical_vars}
\title{Find the categorical variables in a model}
\usage{
find_categorical_vars(fit)
}
\arguments{
\item{fit}{A linear model (fit by \code{\link[=lm]{lm()}}) to refit.}
}
\value{
A character vector of the categorical variables in the model. Note these are not terms,
they are variables, e.g. interactions are not included here, only the variables they are
comprised of.
}
\description{
Find the categorical variables in a model
}
\keyword{internal}
