\name{Fingers.messy}
\docType{data}
\alias{Fingers.messy}
\title{Data from introductory statistics students at a university.}
\description{
  Students at a university taking an introductory statistics course were asked to complete this survey as part of their homework.
}

\usage{Fingers.messy}
\format{
 A dataset with 210 observations on the following 16 variables.
  \describe{
    \item{Sex}{Sex of participant: 1=female, 2=male, 3=prefer not to answer}
    \item{RaceEthnic}{Racial or ethnic background: 1=White, 2=African American, 3=Asian, 4=Latino, 5=Other}
    \item{FamilyMembers}{Members of immediate family (excluding self).}
    \item{SSLast}{Last digit of social security number (NA if no SSN).}
    \item{Year}{Year in school.}
    \item{Job}{Current employment status.}
    \item{MathAnxious}{Agreement with this statement "In general I tend to feel very anxious about mathematics."}
    \item{Interest}{Interest in statistics and the course.}
    \item{GradePredict}{Prediction for final grade in the course from the university's grade points per unit.}
    \item{Thumb}{Length in mm from tip of thumb to the crease between the thumb and palm.}
    \item{Index}{Length in mm from tip of index finger to the crease between the index finger and palm.}
    \item{Middle}{Length in mm from tip of middle finger to the crease between the middle finger and palm.}
    \item{Ring}{Length in mm from tip of ring finger to the crease between the middle finger and palm.}
    \item{Pinkie}{Length in mm from tip of pinkie finger to the crease between the pinkie finger and palm }
    \item{Height}{Height in inches.}
    \item{Weight}{Weight in pounds.}
  }
}
