% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_basis_estimators.R
\name{fit_basis_estimators}
\alias{fit_basis_estimators}
\title{Fit the Zeta Learner: A Highly Flexible Estimator Using Splines or Basis Functions}
\usage{
fit_basis_estimators(
  at,
  a_names,
  z_names,
  w_names,
  outcome,
  outcome_type,
  mediator_type,
  quantile_thresh,
  zeta_learner,
  fold,
  seed
)
}
\arguments{
\item{at}{Training dataframe}

\item{a_names}{Names of treatment variables}

\item{z_names}{Names of mediator  variables}

\item{w_names}{Names of baseline covariate variables}

\item{outcome}{Variable name for the outcome}

\item{outcome_type}{Variable type of the outcome}

\item{mediator_type}{Variable type of the mediator}

\item{quantile_thresh}{Quantile level to set the F-statistic threshold for determining
basis functions for estimation}

\item{zeta_learner}{Stack of algorithms made in SL3 used in ensemble machine
learning to fit Y|A,W}

\item{fold}{Current fold in the cross-validation}

\item{seed}{Seed number for consistent results}
}
\value{
A list of results from fitting the best b-spline model to the data
this list includes the selected learner model, the name of the learner,
the ANOVA fit on the model matrix of the basis functions, the basis used,
and the residual metrics
}
\description{
Identifies variable sets for shift interventions by fitting a flexible estimator
using a discrete Super Learner. This function non-parametrically performs variable importance
of variable sets, such as identifying two mixture components or mixture components and baseline
covariates that explain the outcome given an ANOVA-like decomposition. An F-statistic is used
to threshold the importance of variable sets, and these variable sets are treated as the data-
adaptive parameter.
}
