% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{frm_expand}
\alias{frm_expand}
\title{Expand a formula}
\usage{
frm_expand(frm)
}
\arguments{
\item{frm}{A formula that may have compact terms like \code{a * b}.}
}
\value{
The expanded formula where terms like \code{a * b} are expanded to \code{a + b + a:b}.
}
\description{
Expand a formula
}
\seealso{
formula_building formula_extraction
}
\keyword{internal}
