% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_models.R
\name{formula_string}
\alias{formula_string}
\title{We have to insert spaces where terms were removed from the part model.}
\usage{
formula_string(obj, part, term)
}
\description{
We have to insert spaces where terms were removed from the part model.
}
\keyword{internal}
