% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{frameSetup}
\alias{frameSetup}
\title{frameSetup}
\usage{
frameSetup(anm, x = 0)
}
\arguments{
\item{anm}{\code{\link{anova}} object with no predictors.}

\item{x}{Number of predictiors to be included in the model.}
}
\value{
y An empty data.frame for use by \code{\link{supernova}}.
}
\description{
Sets up a null supernova data.frame
}
