% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{frm_remove_term}
\alias{frm_remove_term}
\title{Remove a term from the right-hand side of a formula}
\usage{
frm_remove_term(frm, term)
}
\arguments{
\item{frm}{The formula to modify.}

\item{term}{The term to drop.}
}
\value{
The formula with the term removed.
}
\description{
Remove a term from the right-hand side of a formula
}
\seealso{
formula_building formula_expansion formula_extraction
}
\keyword{internal}
