% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{frm_string}
\alias{frm_string}
\title{Get the string representation of the formula.}
\usage{
frm_string(frm)
}
\arguments{
\item{frm}{The formula (or formula-coercible object).}
}
\value{
A character string of the formula.
}
\description{
Get the string representation of the formula.
}
\keyword{internal}
