% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_extraction.R
\name{fVal}
\alias{fVal}
\alias{fVal.default}
\alias{fVal.formula}
\title{fVal}
\usage{
fVal(fit, ...)

\method{fVal}{default}(fit, ...)

\method{fVal}{formula}(formula, data = list(), ...)
}
\arguments{
\item{fit}{A \code{\link{lm}} object.}

\item{...}{Passthrough arguments.}

\item{formula}{A \code{\link{formula}}.}

\item{data}{A \code{\link{data.frame}}.}
}
\value{
The F value of the relevant model.
}
\description{
A function to extract F value
}
