% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQValue.R
\name{getQValue}
\alias{getQValue}
\title{Calculates qvalues for any kind of data, failing gracefully. Wrapper for StoreyLab/qvalue function.}
\usage{
getQValue(pvalues)
}
\arguments{
\item{pvalues}{A list of pvalues}
}
\value{
A named list or qobj containing the qvalues at qobj$qvalues, depending on whether the calculation was successful.
}
\description{
Runs the qvalue calculation step
}
