% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridCorMat.R
\name{hybridCorMat}
\alias{hybridCorMat}
\title{Create a hybrid correlation matrix for each subgroup}
\usage{
hybridCorMat(input_file, output_file_prefix, diffCor_cutoff = 0.05,
  corPval_cutoff = 0.05, bufflen = 10000)
}
\arguments{
\item{input_file}{An input filename, pointing to a superNOVA output file.}

\item{output_file_prefix}{The prefix to give to all of the subgroup-specific output files.}

\item{diffCor_cutoff}{The p-value at which to consider a subgroup-specific model versus a global model. Default = 0.05.}

\item{corPval_cutoff}{The p-value at which to consider an edge to have significant correlation between genes versus no correlation. Default = 0.05.}

\item{bufflen}{The number of buffer rows for reading the input file, increasing this will speed up the computation at the expense of memory. Default = 10000.}
}
\value{
None (output to files)
}
\description{
Create a hybrid correlation matrix from a superNOVA output, suitable for downstream MEGENA network creation. Differentially coexpressed edges use a subgroup-specific coexpression model, while others use a global model.
}
\keyword{hybrid}
\keyword{superNOVA}
