% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_row}
\alias{insert_row}
\title{Insert a row of data into a table.}
\usage{
insert_row(df, insert_at, contents)
}
\arguments{
\item{df}{The original data.frame.}

\item{insert_at}{The row in which to insert the data.}

\item{contents}{The row of contents to insert (should be a vector of length
\code{ncol(df)}).}
}
\value{
The original data.frame with the row of data inserted.
}
\description{
Insert a row of data into a table.
}
\keyword{internal}
