% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{insert_rule}
\alias{insert_rule}
\title{Insert a horizontal rule in table for pretty printing}
\usage{
insert_rule(df, insert_at)
}
\arguments{
\item{df}{Original data.frame}

\item{insert_at}{The row in which to insert the new contents (remaining rows will be pushed down).}
}
\value{
df The original data.frame with the new row inserted.
}
\description{
Insert a horizontal rule in table for pretty printing
}
