% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{insertRow}
\alias{insertRow}
\title{insertRow}
\usage{
insertRow(d, nr, rn = NULL)
}
\arguments{
\item{d}{Original data.frame}

\item{nr}{Contents of the new row.}

\item{rn}{The row in which to insert the new contents (remaining rows will be pushed down).}
}
\value{
d The original data.frame with the new row inserted.
}
\description{
Function to insert formatting rows in the output data.frame.
}
