% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrand_q_g.R
\name{integrand_q_g}
\alias{integrand_q_g}
\title{Integrand function for q and g models}
\usage{
integrand_q_g(
  sample_a,
  row_data,
  covars,
  q_model,
  g_model,
  exposure,
  g_delta,
  m_delta,
  upper_bound,
  density_type
)
}
\arguments{
\item{sample_a}{vector of sampled exposure values}

\item{row_data}{dataframe containing the current row data}

\item{covars}{vector of covariate names}

\item{q_model}{sl3 learner object for the q model}

\item{g_model}{sl3 learner object for the g model}

\item{exposure}{character string indicating the exposure variable name}

\item{g_delta}{numeric value for the g model delta}

\item{m_delta}{numeric value for the m model delta}

\item{upper_bound}{numeric value specifying the upper bound for exposure values}

\item{density_type}{character string specifying the density type ("sl" or other)}
}
\value{
output numeric vector of integrand values
}
\description{
Integrand function for q and g models
}
