% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrand_q_r.R
\name{integrand_q_r}
\alias{integrand_q_r}
\title{Integrand Function for Mediation Analysis in SuperNOVA: Works for both Adaptive Quadrature and MC Methods for Integration}
\usage{
integrand_q_r(
  z,
  row_data,
  covars,
  w_names,
  q_model,
  r_model,
  exposure,
  mediator,
  delta,
  upper_a,
  density_type
)
}
\arguments{
\item{z}{Numeric vector, grid points for the mediator.}

\item{row_data}{Data frame, containing the data for a single observation.}

\item{covars}{Character vector, names of the covariate columns in the data.}

\item{w_names}{Character vector, names of the baseline covariate columns in the data.}

\item{q_model}{Fitted Super Learner model, the outcome model given mediator and exposure.}

\item{r_model}{Fitted Super Learner or parametric model, the mediator density given exposure and covariates.}

\item{exposure}{Character, name of the exposure variable in the data.}

\item{mediator}{Character, name of the mediator variable in the data.}

\item{delta}{Numeric, the amount to shift the exposure variable.}

\item{upper_a}{Numeric, the upper bound for the exposure variable.}

\item{density_type}{Character, either "sl" for Super Learner or "parametric" for parametric models.}
}
\value{
Numeric vector, the computed values of the integrand.
}
\description{
Computes the integrand for mediation analysis in SuperNOVA by combining the models for outcome given mediator and exposure (q_model), and the density of the mediator given exposure and covariates (r_model).
}
