% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_m_g.R
\name{integrate_m_g}
\alias{integrate_m_g}
\title{Integrate functions m and g}
\usage{
integrate_m_g(
  av,
  at,
  covars,
  w_names,
  q_model,
  g_model,
  exposure,
  g_delta,
  m_delta
)
}
\arguments{
\item{av}{Validation data}

\item{at}{Training data}

\item{covars}{The mediator variable}

\item{w_names}{Covariate names}

\item{q_model}{A \code{character} vector covariates to adjust for.}

\item{g_model}{The training data}

\item{exposure}{A \code{numeric} indicating the magnitude of the shift to be
computed for the exposure \code{A}. This is passed to the internal}

\item{g_delta}{Amount of delta applied to exposure in g model}

\item{m_delta}{Amount of delta applied to exposure in m model}
}
\value{
A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W).
}
\description{
Integrate functions m and g
}
\details{
Does the double integration as described in lemma 1
}
