% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_psi_aw_g.R
\name{integrate_psi_aw_g}
\alias{integrate_psi_aw_g}
\title{Pseudo regression integration}
\usage{
integrate_psi_aw_g(
  at,
  av,
  covars,
  w_names,
  pseudo_model,
  g_model,
  exposure,
  delta,
  psi_aw,
  n_samples,
  density_type,
  integration_method = "AQ"
)
}
\arguments{
\item{at}{Training data}

\item{av}{Validation data}

\item{covars}{Covariates used in the pseudo model}

\item{w_names}{Baseline covariates used in the g model}

\item{pseudo_model}{Pseudo regression model}

\item{g_model}{The training data}

\item{exposure}{A \code{numeric} indicating the magnitude of the shift to be
computed for the exposure \code{A}. This is passed to the internal}

\item{delta}{Object containing a set of instantiated learners from the
\pkg{sl3}, to be used in fitting an ensemble model.}

\item{psi_aw}{Vector of predictions from pseudo regression model}

\item{n_samples}{Number of MC samples for integration}

\item{density_type}{Type of density estimation used}

\item{integration_method}{Type of integration method to be used}
}
\value{
A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W).
}
\description{
Pseudo regression integration
}
\details{
Does the double integration as described in lemma 1
}
