% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_psi_aw_g_quant.R
\name{integrate_psi_aw_g_quant}
\alias{integrate_psi_aw_g_quant}
\title{Pseudo regression integration for quantized}
\usage{
integrate_psi_aw_g_quant(
  at,
  av,
  covars,
  w_names,
  pseudo_model,
  g_model,
  exposure,
  psi_aw,
  use_multinomial,
  density_type
)
}
\arguments{
\item{at}{Training data}

\item{av}{Validation data}

\item{covars}{Covariates using the m model}

\item{w_names}{Covariates used in the g model}

\item{pseudo_model}{Pseudo regression model}

\item{g_model}{Exposure density/probability model}

\item{exposure}{A \code{numeric} indicating the magnitude of the shift to be
computed for the exposure \code{A}. This is passed to the internal}

\item{psi_aw}{Vector of pseudo regression predictions}

\item{use_multinomial}{TRUE/FALSE whether multinomial regression is used}

\item{density_type}{Type of density estimation sl or hal}

\item{n_bins}{Number of bins in the discrete exposure}
}
\value{
A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W).
}
\description{
Pseudo regression integration for quantized
}
\details{
Does the double integration as described in lemma 1
}
