% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_psi_g.R, R/integrate_psi_g_mc.R
\name{integrate_psi_g}
\alias{integrate_psi_g}
\title{Integrate Psi G for the Da part of the EIF for stochastic mediation}
\usage{
integrate_psi_g(
  av,
  at,
  covars,
  w_names,
  q_model,
  r_model,
  g_model,
  exposure,
  mediator,
  delta,
  n_samples,
  n_iterations,
  density_type,
  integration_method
)

integrate_psi_g(
  av,
  at,
  covars,
  w_names,
  q_model,
  r_model,
  g_model,
  exposure,
  mediator,
  delta,
  n_samples,
  n_iterations,
  density_type,
  integration_method
)
}
\arguments{
\item{covars}{The mediator variable}

\item{w_names}{Covariate names}

\item{q_model}{A \code{character} vector covariates to adjust for.}

\item{r_model}{Mediator density estimator}

\item{g_model}{The training data}

\item{exposure}{A \code{numeric} indicating the magnitude of the shift to be
computed for the exposure \code{A}. This is passed to the internal}

\item{mediator}{The mediator variable name
\code{\link{shift_additive}} and is currently limited to additive shifts.}

\item{delta}{Object containing a set of instantiated learners from the
\pkg{sl3}, to be used in fitting an ensemble model.}

\item{data}{A \code{character} vector of exposures to be shifted.}

\item{bins}{A \code{dataframe} of training data specific to the fold}
}
\value{
A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W).

A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W)
}
\description{
Integrate Psi G for the Da part of the EIF for stochastic mediation

Integrate Psi G for the Da part of the EIF for stochastic mediation using
monte carlo integration
}
\details{
Does the double integration as described in lemma 1

Does the double integration as described in lemma 1
}
