% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_psi_g_cont.R
\name{integrate_psi_g_cont}
\alias{integrate_psi_g_cont}
\title{Integrate Psi G for the Da part of the EIF for stochastic mediation using
Monte Carlo integration}
\usage{
integrate_psi_g_cont(
  av,
  at,
  covars,
  w_names,
  q_model,
  r_model,
  g_model,
  exposure,
  mediator,
  delta,
  n_samples,
  density_type,
  integration_method
)
}
\arguments{
\item{av}{A dataframe with the natural value of the exposure}

\item{at}{A dataframe with the upshift of the exposure}

\item{covars}{A vector of covariate names}

\item{w_names}{A vector of covariate names}

\item{q_model}{A fitted Q-model}

\item{r_model}{A fitted R-model (mediator density estimator)}

\item{g_model}{A fitted G-model (the training data)}

\item{exposure}{A string indicating the name of the exposure variable}

\item{mediator}{A string indicating the name of the mediator variable}

\item{delta}{A numeric indicating the magnitude of the shift for the exposure}

\item{n_samples}{A numeric specifying the number of samples for Monte Carlo integration}

\item{density_type}{A string specifying the type of density estimation ("sl" for Super Learner)}

\item{integration_method}{A string specifying the integration method to use ("MC" for Monte Carlo, "AQ" for Adaptive Quadrature)}
}
\value{
A list with three elements: "d_a" containing the estimates of Da, "phi_aw" containing the
estimates of the outcome mechanism at the natural value of the exposure Q(A, W), and "phi_aw_g" containing the
estimates of the outcome mechanism at the upshift of the exposure Q(A + delta, W)
}
\description{
Integrate Psi G for the Da part of the EIF for stochastic mediation using
Monte Carlo integration
}
\details{
Computes the double integration as described in lemma 1 using
either Monte Carlo or adaptive quadrature methods.
}
