% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_psi_g_discrete.R
\name{integrate_psi_g_discrete}
\alias{integrate_psi_g_discrete}
\title{Integrate Psi G for the Da part of the EIF for stochastic mediation
when A is quartiles}
\usage{
integrate_psi_g_discrete(
  av,
  at,
  covars,
  w_names,
  q_model,
  r_model,
  g_model,
  exposure,
  mediator,
  delta,
  n_bins,
  n_samples = 1000,
  method = "MC",
  mediator_quantized,
  density_type,
  upper_bound,
  use_multinomial
)
}
\arguments{
\item{av}{Validation data}

\item{at}{Training data}

\item{covars}{Covars for the outcome regression model}

\item{w_names}{w_names for A|W}

\item{q_model}{A \code{character} vector covariates to adjust for.}

\item{r_model}{Mediator density estimator}

\item{g_model}{The training data}

\item{exposure}{A \code{numeric} indicating the magnitude of the shift to be
computed for the exposure \code{A}. This is passed to the internal}

\item{mediator}{The mediator variable name
\code{\link{shift_additive}} and is currently limited to additive shifts.}

\item{delta}{Object containing a set of instantiated learners from the
\pkg{sl3}, to be used in fitting an ensemble model.}

\item{n_bins}{Number of bins exposure is discretrized}

\item{n_samples}{Number of samples used in MC integration}

\item{method}{Integration method}

\item{mediator_quantized}{If mediator is discretized}

\item{density_type}{Type of density estimation}

\item{upper_bound}{Upper bound of exposure}

\item{use_multinomial}{if multinomial is used}
}
\value{
A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W)
}
\description{
Integrate Psi G for the Da part of the EIF for stochastic mediation
when A is quartiles
}
\details{
Does the double integration as described in lemma 1
}
