% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_q_g.R
\name{integrate_q_g}
\alias{integrate_q_g}
\title{Integrate functions m and g using monte carlo method}
\usage{
integrate_q_g(
  av,
  at,
  covars,
  w_names,
  q_model,
  g_model,
  exposure,
  g_delta,
  m_delta,
  n_samples,
  density_type,
  lower_bound,
  upper_bound,
  integration_method = "MC"
)
}
\arguments{
\item{covars}{The mediator variable}

\item{w_names}{Covariate names}

\item{q_model}{A \code{character} vector covariates to adjust for.}

\item{g_model}{The training data}

\item{exposure}{A \code{numeric} indicating the magnitude of the shift to be
computed for the exposure \code{A}. This is passed to the internal}

\item{data}{A \code{character} vector of exposures to be shifted.}

\item{delta}{Object containing a set of instantiated learners from the
\pkg{sl3}, to be used in fitting an ensemble model.}
}
\value{
A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W).
}
\description{
Integrate functions m and g using monte carlo method
}
\details{
Does the double integration as described in lemma 1
}
