% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_shift_additive.R
\name{joint_shift_additive}
\alias{joint_shift_additive}
\title{Joint Shift Additive Modified Treatment Policy}
\usage{
joint_shift_additive(
  a1,
  a2,
  w = NULL,
  delta1,
  delta2,
  lower_bound1,
  upper_bound1,
  lower_bound2,
  upper_bound2
)
}
\arguments{
\item{a1}{A \code{numeric} vector of observed treatment values.}

\item{a2}{A \code{numeric} vector of observed treatment values.=}

\item{w}{A \code{numeric} matrix of observed baseline covariate values.}

\item{delta1}{A \code{numeric} indicating the magnitude of the shift to be
computed for the treatment \code{A}.}

\item{delta2}{A \code{numeric} indicating the magnitude of the shift to be
computed for the treatment \code{A}.}

\item{lower_bound1}{Lower bound of expossure}

\item{upper_bound1}{Upper bound of exposure}

\item{lower_bound2}{Lower bound of expossure}

\item{upper_bound2}{Upper bound of exposure}
}
\value{
A \code{numeric} vector containing the shifted exposure values.
}
\description{
Joint Shift Additive Modified Treatment Policy
}
\details{
A simple modified treatment policy that modifes the observed value
of the exposure by shifting it by a value \code{delta}. Note that this
shifting function assumes support of A|W across all strata of W. Same
as shift additive but for two exposures found in the data-adaptive procedure
}
