% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_stoch_shift_est_g_exp.R
\name{joint_stoch_shift_est_g_exp}
\alias{joint_stoch_shift_est_g_exp}
\title{Estimate the Exposure Mechanism via Generalized Propensity Score for Two
Exposure Variable}
\usage{
joint_stoch_shift_est_g_exp(
  exposures,
  deltas,
  g_learner,
  covars,
  av,
  at,
  adaptive_delta,
  hn_trunc_thresh,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{exposures}{A \code{vector} of characters labeling the exposure
variables.}

\item{deltas}{A \code{numeric} value identifying a shift in the observed
value of the exposure under which observations are to be evaluated.}

\item{covars}{A \code{character} labeling the covariate variables}

\item{av}{A \code{dataframe} of validation data specific to the fold}

\item{at}{A \code{dataframe} of training data specific to the fold}

\item{adaptive_delta}{Whether to adaptively change the delta based on positivity
determined from the clever covariate being below the hn_trunc_thresh level}

\item{hn_trunc_thresh}{Truncation level of the clever covariate used in the
adaptive delta method}

\item{pi_learner}{Object containing a set of instantiated learners
from \pkg{sl3}, to be used in fitting an ensemble model.}
}
\value{
A \code{data.table} with four columns, containing estimates of the
generalized propensity score at a downshift (g(A - delta | W)), no shift
(g(A | W)), an upshift (g(A + delta) | W), and an upshift of magnitude two
(g(A + 2 delta) | W).
}
\description{
Estimate the Exposure Mechanism via Generalized Propensity Score for Two
Exposure Variable
}
\details{
Compute the propensity score (exposure mechanism) for the observed
data, including the shift. This gives the propensity score for the observed
data (at the observed A) the counterfactual shifted exposure levels (at
{A - delta}, {A + delta}, and {A + 2 * delta}).
}
