% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_stoch_shift_est_z_exp.R
\name{joint_stoch_shift_est_z_exp}
\alias{joint_stoch_shift_est_z_exp}
\title{Estimate the Exposure Mechanism via Generalized Propensity Score for One
Mediator Variable and Shift an Exposure}
\usage{
joint_stoch_shift_est_z_exp(
  exposures,
  mediator,
  deltas,
  pi_learner,
  w_names,
  a_names,
  z_names,
  av,
  at
)
}
\arguments{
\item{mediator}{Z \code{character} labeling the mediator variable.}

\item{pi_learner}{Object containing a set of instantiated learners
from \pkg{sl3}, to be used in fitting an ensemble model.}

\item{av}{A \code{dataframe} of validation data specific to the fold}

\item{at}{A \code{dataframe} of training data specific to the fold}

\item{exposure}{A \code{character} labeling the exposure variable.}

\item{covars}{A \code{character} labeling the covariate variables}

\item{delta}{A \code{numeric} value identifying a shift in the observed
value of the exposure under which observations are to be evaluated.}
}
\value{
A \code{data.table} with four columns, containing estimates of the
generalized propensity score at a downshift (g(A - delta | W)), no shift
(g(A | W)), an upshift (g(A + delta) | W), and an upshift of magnitude two
(g(A + 2 delta) | W).
}
\description{
Estimate the Exposure Mechanism via Generalized Propensity Score for One
Mediator Variable and Shift an Exposure
}
\details{
Compute the propensity score (mediator mechanism) for the observed
data, including the shift. This gives the propensity score for the observed
data for the mediator (at the observed A) the counterfactual shifted exposure levels (at
{A - delta}, {A + delta}, and {A + 2 * delta}).
}
