% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{level_pairs}
\alias{level_pairs}
\title{Get all pairs for a given vector}
\usage{
level_pairs(levels)
}
\arguments{
\item{levels}{The vector to get pairs for. It is called levels because it was written for the
purpose of comparing levels of a factor to one another with multiple comparisons.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with two columns, group 1 and group 2, where each row is a
unique pair.
}
\description{
The output of this function should match the pairs you get when you run \code{\link[stats:TukeyHSD]{stats::TukeyHSD()}}.
}
\keyword{internal}
