% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_rules_party}
\alias{list_rules_party}
\title{Get rules from partykit object in rule fitting}
\usage{
list_rules_party(x, i = NULL, ...)
}
\arguments{
\item{x}{Partykit glmtree model object}

\item{i}{null}

\item{...}{additional arguments}
}
\value{
List of rules
}
\description{
Get rules from partykit object in rule fitting
}
