% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lower_tri}
\alias{lower_tri}
\title{Find and return the lower triangle of a matrix}
\usage{
lower_tri(x, diag = FALSE)
}
\arguments{
\item{x}{a matrix or other \R object with \code{length(dim(x)) == 2}.
    For back compatibility reasons, when the above is not fulfilled,
    \code{\link[base]{as.matrix}(x)} is called first.}

\item{diag}{logical.  Should the diagonal be included?}
}
\value{
The values in the lower triangular part of the matrix.
}
\description{
Same as \code{\link[=lower.tri]{lower.tri()}} except it returns the values from the matrix (rather than a positional
matrix that lets you look up the values).
}
\keyword{internal}
