% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_learners.R
\name{make_density_superlearner}
\alias{make_density_superlearner}
\title{Make Default Density Estimators}
\usage{
make_density_superlearner()
}
\value{
The density learner used in the g and z density estimation procedure
}
\description{
Make Default Density Estimators
}
