% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_integrand_q_g_r.R
\name{mc_integrand_q_g_r}
\alias{mc_integrand_q_g_r}
\title{Compute the Monte Carlo integrand for double integration in SuperNOVA}
\usage{
mc_integrand_q_g_r(
  sample_a,
  sample_z,
  row_data,
  covars,
  w_names,
  q_model,
  g_model,
  r_model,
  exposure,
  mediator,
  delta,
  upper_a,
  density_type
)
}
\arguments{
\item{sample_a}{A vector of sampled exposure values.}

\item{sample_z}{A vector of sampled mediator values.}

\item{row_data}{A single row of data containing covariates and outcome.}

\item{covars}{A character vector of covariate names.}

\item{w_names}{A character vector of baseline covariate names.}

\item{q_model}{An SL object representing the outcome model.}

\item{g_model}{An SL object representing the density model of exposure given covariates.}

\item{r_model}{An SL object representing the density model of mediator given exposure and covariates.}

\item{exposure}{The name of the exposure variable.}

\item{mediator}{The name of the mediator variable.}

\item{delta}{The exposure shift for the counterfactual calculation.}

\item{upper_a}{The upper bound for the exposure.}

\item{density_type}{A string indicating the density estimation method, either "sl" or "nonparametric".}
}
\value{
A numeric vector representing the Monte Carlo integrand values for the given sample points.
}
\description{
This function computes the Monte Carlo integrand for double integration in the SuperNOVA package.
It combines the models for the outcome given mediator and exposure (q_model),
the density of the mediator given exposure and covariates (r_model),
and the density of the exposure given covariates (g_model).
}
