% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{new_pairwise_comparisons_tbl}
\alias{new_pairwise_comparisons_tbl}
\title{Constructor for pairwise comparison tables}
\usage{
new_pairwise_comparisons_tbl(tbl, term, fit, fwer, alpha, correction)
}
\arguments{
\item{tbl}{A \code{\link{tibble}}-like object.}

\item{term}{The term the table describes.}

\item{fit}{The linear model the term comes from.}

\item{fwer}{The family-wise error-rate for the group of tests in the table.}

\item{alpha}{The alpha to use when computing the family-wise error-rate.}

\item{correction}{The type of alpha correction the tests in the table use.}
}
\value{
A tibble subclassed as \code{pairwise_comparison_tbl}. These have custom printers and retain
their attributes when subsetted.
}
\description{
Constructor for pairwise comparison tables
}
\keyword{internal}
