% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIEHS_data1.R
\docType{data}
\name{NIEHS_data_1}
\alias{NIEHS_data_1}
\title{Data 1 from the NIEHS mixtures workshop}
\format{
A data frame with 500 rows and 9 variables:
}
\source{
\url{https://github.com/niehs-prime/2015-NIEHS-MIxtures-Workshop}
}
\usage{
NIEHS_data_1
}
\description{
A dataset containing simulated mixture data from the NIEHS
}
\keyword{datasets}
