% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Pad x to length of y}
\usage{
pad(x, y, after = length(x), pad = NA)
}
\arguments{
\item{x}{The vector to pad.}

\item{y}{The vector with target length.}

\item{after}{A subscript, after which the padding is to be appended.}

\item{pad}{The value to pad the vector with.}
}
\value{
The padded vector.
}
\description{
Pad x to length of y
}
