% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_len}
\alias{pad_len}
\title{Pad x to a given output length}
\usage{
pad_len(x, output_length, after = length(x), pad = NA)
}
\arguments{
\item{x}{The vector to pad.}

\item{output_length}{The length to pad the vector to.}

\item{after}{A subscript, after which the padding is to be appended.}

\item{pad}{The value to pad the vector with.}
}
\value{
The padded vector.
}
\description{
Pad x to a given output length
}
\keyword{internal}
