% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_line}
\alias{paste_line}
\title{Paste together lines of text.}
\usage{
paste_line(...)
}
\arguments{
\item{...}{Vectors of lines of text.}
}
\value{
Check out the \link{paste} function for more information.
}
\description{
The lines are joined together with a newline (\verb{\\n}) character.
}
\keyword{internal}
