% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_results.R
\name{plot_sim_results}
\alias{plot_sim_results}
\title{Plot ROC curves for superNOVA simulated data observed and analysed results for the two group case.}
\usage{
plot_sim_results(exp_vs_obs_table, method = "superNOVA",
  samples = "unknown", output = T)
}
\arguments{
\item{exp_vs_obs_table}{The expected versus observed table from analyse_sim_results().}

\item{method}{The name of the differential coexpression method on the plot. Default=superNOVA}

\item{samples}{The number of samples that were analyzed to generate the results.}

\item{output}{Should a file with the ROC curve be written?}
}
\value{
The ROC curves for all 6 differentially coexpressed classes (excludes null classes).
}
\description{
Provides a way to plot ROC curves for the two group case. This will not work for 3 or more simulated groups.
}
\keyword{plot}
\keyword{simulate}
\keyword{superNOVA}
