% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_extraction.R
\name{PRE}
\alias{PRE}
\alias{PRE.default}
\alias{PRE.formula}
\title{PRE}
\usage{
PRE(fit, ...)

\method{PRE}{default}(fit, ...)

\method{PRE}{formula}(formula, data = list(), ...)
}
\arguments{
\item{fit}{A \code{\link{lm}} object.}

\item{...}{Passthrough arguments.}

\item{formula}{A \code{\link{formula}}.}

\item{data}{A \code{\link{data.frame}}.}
}
\value{
The PRE value of the relevant model.
}
\description{
A function to extract PRE values
}
