% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lm.R
\name{print.lm}
\alias{print.lm}
\title{Print the output of lm() with fitted equation.}
\usage{
\method{print}{lm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{The fitted linear model to print.}

\item{digits}{The minimal number of significant digits.}

\item{...}{Ignored.}
}
\value{
Invisibly return the fitted linear model.
}
\description{
Print the output of lm() with fitted equation.
}
