% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.supernova}
\alias{print.supernova}
\title{print.supernova}
\usage{
\method{print}{supernova}(x, pcut = 4, ...)
}
\arguments{
\item{x}{A \code{\link{supernova}} object.}

\item{pcut}{The integer number of decimal places of p-values to show.}

\item{...}{Additional display arguments.}
}
\description{
A print method for the supernova class
}
