% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{printHelp}
\alias{printHelp}
\title{printHelp}
\usage{
printHelp(x, digits = max(3L, getOption("digits") - 2L), dig.tst = max(1L,
  min(5L, digits - 1L)), cs.ind = 1:k, tst.ind = k + 1, na.print = "",
  ...)
}
\arguments{
\item{x}{An atomic character.}

\item{digits}{From \code{\link{options}} by default.}

\item{dig.tst}{As in printCoefMat.}

\item{cs.ind}{As in printCoefMat.}

\item{tst.ind}{As in printCoefMat.}

\item{na.print}{Alternate string to print for NA values.}

\item{...}{Additional arguments.}
}
\value{
A character with correct numeric formatting.
}
\description{
A function to help with formatting in \code{\link{print.supernova}}
}
