% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_integrand_q_g_r.R
\name{quad_integrand_q_g_r}
\alias{quad_integrand_q_g_r}
\title{Integrand Function for Q, g, r Model Adaptive Quadrature Double Integration in SuperNOVA}
\usage{
quad_integrand_q_g_r(
  a,
  z,
  row_data,
  covars,
  w_names,
  q_model,
  g_model,
  r_model,
  exposure,
  mediator,
  delta,
  upper_a,
  density_type
)
}
\arguments{
\item{a}{Numeric matrix, grid points for the exposure.}

\item{z}{Numeric matrix, grid points for the mediator.}

\item{row_data}{Data frame, containing the data for a single observation.}

\item{covars}{Character vector, names of the covariate columns in the data.}

\item{w_names}{Character vector, names of the baseline covariate columns in the data.}

\item{q_model}{Fitted Super Learner model, the outcome model given mediator and exposure.}

\item{g_model}{Fitted Super Learner or parametric model, the exposure density given covariates.}

\item{r_model}{Fitted Super Learner or parametric model, the mediator density given exposure and covariates.}

\item{exposure}{Character, name of the exposure variable in the data.}

\item{mediator}{Character, name of the mediator variable in the data.}

\item{delta}{Numeric, the amount to shift the exposure variable.}

\item{upper_a}{Numeric, the upper bound for the exposure variable.}

\item{density_type}{Character, either "sl" for Super Learner or "hal" for highly adaptive lasso}
}
\value{
Numeric matrix, the computed values of the integrand.
}
\description{
Computes the integrand for adaptive quadrature double integration in SuperNOVA by combining the models for outcome given mediator and exposure (q_model), the density of the mediator given exposure and covariates (r_model), and the density of the exposure given covariates (g_model).
}
