% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{refit_categorical}
\alias{refit_categorical}
\title{Refit a model, dropping any non-categorical terms.}
\usage{
refit_categorical(fit)
}
\arguments{
\item{fit}{A linear model (fit by \code{\link[=lm]{lm()}}) to refit.}
}
\value{
A linear model that only has categorical predictors.
}
\description{
Refit a model, dropping any non-categorical terms.
}
\keyword{internal}
