% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename}
\alias{rename}
\title{Rename a column in a data frame}
\usage{
rename(data, col_name, replacement)
}
\arguments{
\item{data}{A data frame to modify.}

\item{col_name}{A character vector of columns to rename.}

\item{replacement}{A character vector of replacement column names.}
}
\value{
Returns the renamed data frame.
}
\description{
Rename a column in a data frame
}
\keyword{internal}
