% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_type}
\alias{resolve_type}
\title{Convert SS type parameter to the corresponding numeric value}
\usage{
resolve_type(type)
}
\arguments{
\item{type}{The value to convert, either string or numeric.}
}
\value{
The numeric value corresponding to the input.
}
\description{
Convert SS type parameter to the corresponding numeric value
}
\keyword{internal}
