% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supernova.R
\name{row_error}
\alias{row_error}
\title{Compute and construct an ANOVA table row for an error term}
\usage{
row_error(term, description, fit)
}
\arguments{
\item{term}{The name of the term the row describes (e.g. Error or Total).}

\item{description}{An optional, short description of the term (pedagogical).}

\item{fit}{The model we are describing error from.}
}
\value{
A \code{\link{tibble_row}} with the properties initialized. The code has been written to be as
simple and understandable as possible. Please take a look at the source and offer any
suggestions for improvement!
}
\description{
Compute and construct an ANOVA table row for an error term
}
\keyword{internal}
