% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supernova.R
\name{row_term}
\alias{row_term}
\title{Compute and construct an ANOVA table row for a term.}
\usage{
row_term(name, description, models, term)
}
\arguments{
\item{description}{An optional, short description of the term (pedagogical).}

\item{models}{The models created by \code{\link[=generate_models]{generate_models()}} for comparison.}

\item{term}{The term to compute the row for.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble_row()}} with the properties initialized. The code has been written to
be as understandable as possible. Please take a look at the source and offer any suggestions
for improvement!
}
\description{
"Term" is loosely defined here and is probably better understood as "everything in the table that
is not an error row.
}
\keyword{internal}
