% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{select_terms}
\alias{select_terms}
\title{Select terms based on the user's \code{term} specification}
\usage{
select_terms(fit, term = NULL)
}
\arguments{
\item{fit}{A model fit by \code{\link[=lm]{lm()}} or \code{\link[=aov]{aov()}} (or similar).}

\item{term}{If \code{NULL}, use each categorical term in the model. Otherwise, only use the given
term.}
}
\value{
A character vector of terms to run analyses on.
}
\description{
Before returning the selection, ensure that the term we are subsetting on exists.
}
\keyword{internal}
