% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Servers}
\alias{Servers}
\title{Servers data}
\format{
A data frame with 44 observations on the following 2 variables.
\describe{
\item{\code{ServerID}}{A number assigned to each server.}
\item{\code{Tip}}{How much the tip was.}
}
}
\usage{
Servers
}
\description{
Data about tips collected from an experiment with 44 servers at a restaurant.
}
\details{
Note: these data will be removed in future versions in favor of \code{\link{Tables}}.
}
\keyword{datasets}
