% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_additive_joint.R
\name{shift_additive_joint}
\alias{shift_additive_joint}
\title{Simple Additive Modified Treatment Policy for Two Exposures}
\usage{
shift_additive_joint(a, w = NULL, deltas)
}
\arguments{
\item{deltas}{A \code{numeric} list indicating the magnitude of the shift to be
computed for the treatment \code{A}.}

\item{A}{A \code{numeric} vector of observed treatment values.}

\item{W}{A \code{numeric} matrix of observed baseline covariate values.}
}
\value{
A \code{numeric} vector containing the shifted exposure values.
}
\description{
Simple Additive Modified Treatment Policy for Two Exposures
}
\details{
A simple modified treatment policy that modifes the observed value
of the exposure by shifting it by a value \code{delta}. Note that this
shifting function assumes support of A|W across all strata of W. This function
shifts two treatments or exposures.
}
