% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Title}
\usage{
simulate_data(
  n_obs = 2000,
  mu = c(2.5, 3, 4),
  sigma_mod = matrix(c(1, 0.5, 0.8, 0.5, 1, 0.7, 0.8, 0.7, 1), nrow = 3, ncol = 3)
)
}
\arguments{
\item{sigma_mod}{}
}
\value{

}
\description{
Title
}
