% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mediation_data.R
\name{simulate_mediation_data}
\alias{simulate_mediation_data}
\title{DGP for testing SuperNOVA with mediation}
\usage{
simulate_mediation_data(n_obs = 1e+05, delta = 1, n_bins = 10)
}
\arguments{
\item{n_obs}{Number of observations}

\item{delta}{Delta for M1 Exposure}

\item{n_bins}{Number of bins to discretize exposure}
}
\value{
A data frame of simulated mediation data
}
\description{
DGP for testing SuperNOVA with mediation
}
