% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snMEGENA.R
\name{snMEGENA}
\alias{snMEGENA}
\title{SuperNOVA subgroup-specific hybrid MEGENA networks}
\usage{
snMEGENA(data.file, subtype.name, module.pval = 0.05, hub.pval = 0.05,
  cor.perm = 10, hub.perm = 50, annot.table = TRUE, n.cores = 1)
}
\arguments{
\item{data.file}{Path to the hybrid correlation matrix.}

\item{subtype.name}{Name of the subgroup for the output.}

\item{module.pval}{Module significance p-value. Recommended is 0.05 (default).}

\item{hub.pval}{Connectivity significance p-value based random tetrahedral networks. Recommended is 0.05 (default).}

\item{cor.perm}{Number of permutations for calculating FDRs for all correlation pairs. Default is 10.}

\item{hub.perm}{Number of permutations for calculating connectivity significance p-value. Default is 50.}

\item{annot.table}{Should we create an annotation table for the output? Default = TRUE.}

\item{n.cores}{Number of cores to use with MEGENA. Default is 1 (single-threaded).}
}
\value{
None (output to files)
}
\description{
Takes in an subgroup-specific hybrid correlation matrix from hybridCorMat and builds a MEGENA network from it.
}
\keyword{MEGENA}
\keyword{hybrid}
\keyword{superNOVA}
