% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sub_matches}
\alias{sub_matches}
\title{Replace and subset only matched elements}
\usage{
sub_matches(x, pattern, replacement)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. \link{Long vectors} are
supported.}

\item{pattern}{character string containing a \link{regular expression} (or
character string for \code{fixed = TRUE}) to be matched in the given
character vector. Coerced by \code{\link{as.character}} to a character
string if possible. If a character vector of length 2 or more is supplied,
the first element is used with a warning.  Missing values are allowed
except for \code{regexpr}, \code{gregexpr} and \code{regexec}.}

\item{replacement}{a replacement for matched pattern in \code{sub} and
\code{gsub}. Coerced to character if possible. For \code{fixed = FALSE}
this can include backreferences \code{"\\1"} to \code{"\\9"} to
parenthesized subexpressions of \code{pattern}.  For \code{perl = TRUE}
only, it can also contain \code{"\\U"} or \code{"\\L"} to convert the rest
of the replacement to upper or lower case and \code{"\\E"} to end case
conversion. If a character vector of length 2 or more is supplied, the
first element is used with a warning.  If \code{NA}, all elements in the
result corresponding to matches will be set to \code{NA}.}
}
\value{
a character vector containing only the modified elements.
}
\description{
Search and subset on pattern matches, then make replacements.
}
\keyword{internal}
