% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{superanova}
\alias{superanova}
\title{superanova}
\usage{
superanova(x)
}
\arguments{
\item{x}{A \code{\link{lm}} object.}
}
\value{
A \code{\link{supernova}} object.
}
\description{
Alias for supernova
}
