% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superNOVAFindSignificant.R
\name{superNOVAFindSignificant}
\alias{superNOVAFindSignificant}
\title{Find groups of differentially coexpressed gene symbols.}
\usage{
superNOVAFindSignificant(superNOVA_res, pval_gene_thresh = 0.05,
  adjusted = FALSE, classes = TRUE, geneNameCol = c("Gene1",
  "Gene2"), unique_genes = FALSE, compare = NULL, bySlope = FALSE)
}
\arguments{
\item{superNOVA_res}{The table of differential correlations outputted from superNOVA. Expected to have pValDiff or qValDiff columns as well as group_order, groupCor, globalCor, groupSlope, globalSlope, Gene1, Classes columns.}

\item{pval_gene_thresh}{p-value threshold to call a gene as having significant differential correlation or not. Default = 0.05}

\item{adjusted}{Logical indicating whether adjusted p-values from the differential correlation table (i.e., column "qValDiff", when adjusted = TRUE) or unadjusted p-values (i.e., column "pValDiff", when adjusted = FALSE) should be used to subset the table into significant and non-significant portions. Default = FALSE}

\item{classes}{Logical indicator specifying whether individual differential correlation gene classes should be extracted from the table or not. If not, compare MUST be set. Default = TRUE}

\item{geneNameCol}{Character vector specifying the name of the columns that are used to extract the gene symbols. Note that the default is c("Gene1", "Gene2"), but this only makes sense in the context of a full DGCA experiment. In the case of a splitSet, you may want to use "Gene1" to avoid counting the splitSet names in all of the categories.}

\item{unique_genes}{Logical, if TRUE indicates that unique gene symbols from each category compared to the other groups should be chosen prior to GO enrichment analysis.}

\item{compare}{Names of one or two groups to compare. If set to one group name, compare that group to the global model. If set to two group names, pairwise compare with each other. Default = NULL.}

\item{bySlope}{Whether to use the slope instead of the correlation rho when comparing one or two groups. This parameter has no effect when classes=TRUE.}
}
\value{
A list of significantly differentially coexpressed genes.
}
\description{
Takes a table of differentially correlated genes with respect to one gene in the Gene2 column and returns the a list of vectors with unique, non-NA gene symbols for genes in each of the differentially correlated classes.
}
