\name{Survey}
\docType{data}
\alias{Survey}
\title{Students at a university were asked to enter a random number between 1-20 into a survey.}
\description{
  Students at a university taking an introductory statistics course were asked to complete this survey as part of their homework.
}

\usage{Survey}
\format{
 A dataset with 211 observations on the following 1 variable:
  \describe{
    \item{Any1_20}{The random number between 1 and 20 that a student thought of.}
  }
}
