% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tables}
\alias{Tables}
\title{Tables data}
\format{
A data frame with 44 observations on the following 2 variables.
\describe{
\item{TableID}{A number assigned to each table.}
\item{Tip}{How much the tip was.}
}
}
\usage{
Tables
}
\description{
Data about tips collected from an experiment with 44 tables at a restaurant.
}
\keyword{datasets}
