% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TipExperiment}
\alias{TipExperiment}
\title{Data from an experiment about smiley faces and tips}
\format{
A data frame with 44 observations on the following 3 variables.
\describe{
\item{TableID}{A number assigned to each table.}
\item{Tip}{How much the tip was.}
\item{Condition}{Which experimental condition the table was randomly assigned to.}
}
}
\usage{
TipExperiment
}
\description{
Tables were randomly assigned to receive checks that either included or did not include a drawing
of a smiley face. Data was collected from 44 tables in an effort to examine whether the added
smiley face would cause more generous tipping.
}
\keyword{datasets}
