% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_exposhift.R
\name{tmle_exposhift}
\alias{tmle_exposhift}
\title{Targeted Minimum Loss Estimate of Counterfactual Mean of Stochastic Shift
Intervention}
\usage{
tmle_exposhift(
  data_internal,
  delta,
  Qn_scaled,
  Qn_unscaled,
  Hn,
  fluctuation = c("standard", "weighted"),
  eif_reg_type = c("hal", "glm"),
  y,
  estimator = "tmle"
)
}
\arguments{
\item{data_internal}{A \code{data.table} constructed internally by a call to
expo_shift. This contains most of the data for computing the
targeted minimum loss (TML) estimator.}

\item{delta}{A \code{numeric} value indicating the shift in the treatment to
be used in defining the target parameter. This is defined with respect to
the scale of the treatment (A).}

\item{Qn_scaled}{An object providing the value of the outcome evaluated after
imposing a shift in the treatment. This object is passed in after being
constructed by a call to the internal function for Q estimation.}

\item{Hn}{An object providing values of the auxiliary ("clever")
covariate, constructed from the treatment mechanism and required for
targeted minimum loss-based estimation. This object object should be passed
in after being constructed by a call to clever covariate construction.}

\item{fluctuation}{The method to be used in the submodel fluctuation step
(targeting step) to compute the TML estimator. The choices are "standard"
and "weighted" for where to place the auxiliary covariate in the logistic
tilting regression.}

\item{eif_reg_type}{Whether a flexible nonparametric function ought to be
used in the dimension-reduced nuisance regression of the targeting step for
the censored data case. By default, the method used is a nonparametric
regression based on the Highly Adaptive Lasso.
Set this to \code{"glm"} to instead use a simple linear regression model.
In this step, the efficient influence function (EIF) is regressed against
covariates contributing to the censoring mechanism (i.e., EIF ~ V | C = 1).}

\item{y}{Outcome variable}
}
\value{
S3 object of class \code{txshift} containing the results of the
procedure to compute a TML estimate of the treatment shift parameter.
}
\description{
Targeted Minimum Loss Estimate of Counterfactual Mean of Stochastic Shift
Intervention
}
\details{
Invokes the procedure to construct a targeted minimum loss estimate
(TMLE) of the counterfactual mean under a modified treatment policy.
}
