% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{update}
\alias{update}
\title{Robust version of \code{\link[stats]{update}}}
\usage{
update(old, new, ...)
}
\arguments{
\item{old}{An existing fit from a model function such as \code{\link{lm}}, 
\code{\link{glm}} and many others.}

\item{new}{Changes to the formula; see \code{\link{update.formula}} for details.}

\item{...}{Additional arguments to the call, or arguments with changed values. 
Use name = NULL to remove the argument name.}
}
\value{
If evaluate = TRUE the fitted object, otherwise the updated call.
}
\description{
Robust version of \code{\link[stats]{update}}
}
