% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_in_env}
\alias{update_in_env}
\title{Update a model in the environment the model was created in}
\usage{
update_in_env(object, formula., ...)
}
\arguments{
\item{formula.}{Changes to the formula -- see \code{update.formula} for
   details.}

\item{...}{Additional arguments to the call, or arguments with
   changed values. Use \code{name = NULL} to remove the argument \code{name}.}
}
\value{
The updated model is returned.
}
\description{
\code{\link[stats:update]{stats::update()}} will perform the update in \code{\link[=parent.frame]{parent.frame()}} by default, but this can cause
problems when the update is called by another function (so the parent frame is no longer the
environment the user is in).
}
