% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{variables}
\alias{variables}
\title{Extract the variables from a model}
\usage{
variables(object)
}
\arguments{
\item{object}{An \code{\link{lm}} or \code{\link{supernova}} object}
}
\value{
A list containing the \code{outcome} and \code{predictor} variables
  in the model.
}
\description{
Extract the variables from a model
}
