% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anova_tbl}
\alias{anova_tbl}
\title{ANOVA table with nicer column names.}
\usage{
anova_tbl(model)
}
\arguments{
\item{model}{A model fit by \code{\link[=lm]{lm()}} or \code{lme4::lmer()}}
}
\value{
An ANOVA table with standard column names.
}
\description{
ANOVA table with nicer column names.
}
\keyword{internal}
