% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_CIs}
\alias{calc_CIs}
\title{Calculate Confidence Interals}
\usage{
calc_CIs(psi, psi_se)
}
\description{
Gives information on which variable sets are used in the
basis function in the data-adaptive estimation section
}
