% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_utils.R
\name{calc_final_effect_mod_param}
\alias{calc_final_effect_mod_param}
\title{Calculates the Effect Modification Shift Parameter}
\usage{
calc_final_effect_mod_param(
  tmle_fit_av,
  tmle_fit_at,
  exposure,
  at,
  av,
  effect_m_name,
  fold_k,
  em_learner
)
}
\description{
Use a decision tree estimator to regress the eif of an exposure
on to a covariate - both of these have been data adaptively determined.
}
