% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_utils.R
\name{calc_final_effect_mod_param}
\alias{calc_final_effect_mod_param}
\title{Calculates the Effect Modification Shift Parameter}
\usage{
calc_final_effect_mod_param(
  tmle_fit_av,
  tmle_fit_at,
  exposure,
  at,
  av,
  effect_m_name,
  fold_k,
  em_learner
)
}
\arguments{
\item{tmle_fit_av}{TMLE results for the validation fold.}

\item{tmle_fit_at}{TMLE results for the training fold.}

\item{exposure}{The identified exposure variable as a \code{character} string.}

\item{at}{Training dataset.}

\item{av}{Validation dataset.}

\item{effect_m_name}{The name of the effect modifier variable as a \code{character} string.}

\item{fold_k}{The fold in which the effect modification was found.}

\item{em_learner}{A list of learners for the Super Learner, specifically for decision trees.}
}
\description{
This function uses a decision tree estimator to regress the difference in expected outcomes
of an exposure onto a covariate. Both the exposure and the covariate have been determined through data-adaptive methods.
If no rules are found, a median split is applied.
}
