% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_utils.R
\name{calc_final_ind_shift_param}
\alias{calc_final_ind_shift_param}
\title{Calculates the Individual Shift Parameter}
\usage{
calc_final_ind_shift_param(tmle_fit, exposure, fold_k)
}
\arguments{
\item{tmle_fit}{TMLE results for the individual shift}

\item{exposure}{The exposure identified}

\item{fold_k}{Fold exposure is being shifted in}
}
\description{
Simply subtract the psi estimates of a shift compared to no
shift. Use the delta method to do the same thing for the eif to derive
variance estimates for this new parameter.
}
