% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_utils.R
\name{calc_final_joint_shift_param}
\alias{calc_final_joint_shift_param}
\title{Calculates the Joint Shift Parameter}
\usage{
calc_final_joint_shift_param(joint_shift_fold_results, exposures, fold_k)
}
\description{
Estimates the shift parameter for a joint shift
}
