% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_joint_results}
\alias{calc_joint_results}
\title{Calculate the Joint Parameter}
\usage{
calc_joint_results(results_element)
}
\arguments{
\item{results_element}{Joint results for formatting}
}
\description{
Using the output results for the eif use the delta method
to simply subtract the joint estimate from the no shift estimate. Do
the same for the eif to get the variance for this new parameter.
}
