\name{Fingers}
\docType{data}
\alias{Fingers}
\title{Data from introductory statistics students at a university.}
\description{
  Students at a university taking an introductory statistics course were asked to complete this survey as part of their homework.
}

\usage{Fingers}
\format{
 A dataset with 157 observations on the following 16 variables.
  \describe{
    \item{Sex}{Sex of participant.}
    \item{RaceEthnic}{Racial or ethnic background.}
    \item{FamilyMembers}{Members of immediate family (excluding self).}
    \item{SSLast}{Last digit of social security number (NA if no SSN).}
    \item{Year}{Year in school: 1=First, 2=Second, 3=Third, 4=Fourth, 5=Other}
    \item{Job}{Current employment status: 0=not working, 1=part-time job, 2=full-time job}
    \item{MathAnxious}{Agreement with this statement "In general I tend to feel very anxious about mathematics": 2=Strongly Agree, 1=Agree, 0=Neither Agree nor Disagree, -1=Disagree, -2=Strongly Disagree}
    \item{Interest}{Interest in statistics and the course: 2=very interested in course and statistics, 1=somewhat interested, 0=no interest, -1=dread the course.}
    \item{GradePredict}{Prediction for final grade in the course from the university's grade points per unit: 4.0=A, 3.7=A-, 3.3=B+, 3.0=B, 2.7=B-, 2.3=C+, 2.0=C, 1.7=C-, 1.3=Below C- }
    \item{Thumb}{Length in mm from tip of thumb to the crease between the thumb and palm.}
    \item{Index}{Length in mm from tip of index finger to the crease between the index finger and palm.}
    \item{Middle}{Length in mm from tip of middle finger to the crease between the middle finger and palm.}
    \item{Ring}{Length in mm from tip of ring finger to the crease between the middle finger and palm.}
    \item{Pinkie}{Length in mm from tip of pinkie finger to the crease between the pinkie finger and palm }
    \item{Height}{Height in inches.}
    \item{Weight}{Weight in pounds.}
  }
}
