% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_basis_estimators.R
\name{fit_basis_estimators}
\alias{fit_basis_estimators}
\title{Fit the zeta learner, a highly flexible estimator using splines
or basis functions in order to identify variable sets for the shift
interventions.}
\usage{
fit_basis_estimators(
  at,
  covars,
  outcome,
  family,
  quantile_thresh,
  zeta_learner,
  fold,
  seed
)
}
\arguments{
\item{at}{Training dataframe}

\item{covars}{Covariates to be used as predictors in the Super Learner}

\item{outcome}{Variable name for the outcome}

\item{zeta_learner}{Stack of algorithms made in SL 3 used in ensemble machine
learning to fit Y|A,W}

\item{fold}{Current fold in the cross-validation}

\item{seed}{Seed number for consistent results}

\item{max_iter}{Max number of iterations of iterative backfitting algorithm}

\item{verbose}{Run in verbose setting}
}
\value{
Rules object. TODO: add more detail here.
}
\description{
Fit the zeta learner, a highly flexible estimator using splines
or basis functions in order to identify variable sets for the shift
interventions.
}
\details{
Finds the best fitting flexible estimator using a discrete
Super Learner. This function non-parametrically does variable importance
of variable sets, such as identificaiton of two mixture components or
mixture components and baseline covariates which explain the outcome given
an ANOVA like decomposition. We use an F-statistic to threshold the
importance of variable sets and treat these variable sets as the data-
adaptive parameter.
}
