% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenChow.R
\name{flattenChow}
\alias{flattenChow}
\title{Convert the output from superNOVA chowCor to an annotated edgelist, suitable for writing out to a .tsv file.}
\usage{
flattenChow(chow_result)
}
\arguments{
\item{chow_result}{A list of matrices as output by chowCor. Required}
}
\value{
Returns a data.frame, sorted by superNOVA p-value, where each row is a feature pair, with columns "Gene1","Gene2","groupCor","groupCorPval","globalCor","globalCorP","pValDiff","Classes".
}
\description{
Takes in results from chowCor as a list of 2d matrices and converts it into a single data.frame, where the first two columns are an edgelist of all pairs compared.
}
\keyword{superNOVA}
