% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{frm_build}
\alias{frm_build}
\title{Build a formula from terms}
\usage{
frm_build(lhs, rhs)
}
\arguments{
\item{lhs}{The outcome term for the left-hand side.}

\item{rhs}{The terms for the right-hand side.}
}
\value{
The right-hand side terms are joined with \code{+}. Then, the right-hand side is joined
  to the left and returned as a \code{\link{formula}}.
}
\description{
Build a formula from terms
}
\seealso{
formula_extraction formula_expansion
}
\keyword{internal}
