% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_stoch_shift_est_g_exp.R
\name{indiv_stoch_shift_est_g_exp}
\alias{indiv_stoch_shift_est_g_exp}
\title{Estimate the Exposure Mechanism via Generalized Propensity Score for One
Exposure Variable}
\usage{
indiv_stoch_shift_est_g_exp(
  exposure,
  delta,
  pi_learner,
  covars,
  av,
  at,
  adaptive_delta,
  hn_trunc_thresh
)
}
\arguments{
\item{exposure}{A \code{character} labeling the exposure variable.}

\item{delta}{A \code{numeric} value identifying a shift in the observed
value of the exposure under which observations are to be evaluated.}

\item{pi_learner}{Object containing a set of instantiated learners
from \pkg{sl3}, to be used in fitting an ensemble model.}

\item{covars}{A \code{character} labeling the covariate variables}

\item{av}{A \code{dataframe} of validation data specific to the fold}

\item{at}{A \code{dataframe} of training data specific to the fold}

\item{adaptive_delta}{Whether or not to adaptively adjust delta based on
positivity (estimated from the clever covariate) meeting the hn_trunc_thresh
level}

\item{hn_trunc_thresh}{Level of the clever covariate in the adaptive delta
procedure}
}
\value{
A \code{data.table} with four columns, containing estimates of the
generalized propensity score at a downshift (g(A - delta | W)), no shift
(g(A | W)), an upshift (g(A + delta) | W), and an upshift of magnitude two
(g(A + 2 delta) | W).
}
\description{
Estimate the Exposure Mechanism via Generalized Propensity Score for One
Exposure Variable
}
\details{
Compute the propensity score (exposure mechanism) for the observed
data, including the shift. This gives the propensity score for the observed
data (at the observed A) the counterfactual shifted exposure levels (at
{A - delta}, {A + delta}, and {A + 2 * delta}).
}
