% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_rule}
\alias{insert_rule}
\title{Insert a horizontal rule in a table for pretty printing}
\usage{
insert_rule(df, insert_at)
}
\arguments{
\item{df}{The original data.frame}

\item{insert_at}{The row in which to insert the dashes.}
}
\value{
The original data.frame with the horizontal rule inserted.
}
\description{
Insert a horizontal rule in a table for pretty printing
}
\keyword{internal}
