% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_psi_aw_g.R
\name{integrate_psi_aw_g}
\alias{integrate_psi_aw_g}
\title{Pseudo regression integration}
\usage{
integrate_psi_aw_g(
  data,
  covars,
  w_names,
  pseudo_model,
  g_model,
  exposure,
  delta,
  psi_aw,
  pseudo_reg_backscale
)
}
\arguments{
\item{data}{A \code{character} vector of exposures to be shifted.}

\item{covars}{The mediator variable}

\item{w_names}{Covariate names}

\item{g_model}{The training data}

\item{exposure}{A \code{numeric} indicating the magnitude of the shift to be
computed for the exposure \code{A}. This is passed to the internal}

\item{delta}{Object containing a set of instantiated learners from the
\pkg{sl3}, to be used in fitting an ensemble model.}

\item{q_model}{A \code{character} vector covariates to adjust for.}
}
\value{
A \code{data.table} with two columns, containing estimates of the
outcome mechanism at the natural value of the exposure Q(A, W) and an
upshift of the exposure Q(A + delta, W).
}
\description{
Pseudo regression integration
}
\details{
Does the double integration as described in lemma 1
}
