% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise}
\alias{pairwise}
\alias{pairwise_t}
\alias{pairwise_bonferroni}
\alias{pairwise_tukey}
\title{Compute all pairwise comparisons between category levels}
\usage{
pairwise(fit, correction = "Tukey", term = NULL, alpha = 0.05, plot = FALSE)

pairwise_t(fit, term = NULL, alpha = 0.05, correction = "none")

pairwise_bonferroni(fit, term = NULL, alpha = 0.05)

pairwise_tukey(fit, term = NULL, alpha = 0.05)
}
\arguments{
\item{fit}{A model fit by \code{\link[=lm]{lm()}} or \code{\link[=aov]{aov()}} (or similar).}

\item{correction}{The type of correction (if any) to perform to maintain the family-wise
error-rate specified by \code{alpha}: - \strong{Tukey}: computes Tukey's Honestly Significant Differences
(see \code{\link[=TukeyHSD]{TukeyHSD()}}) - \strong{Bonferroni}: computes pairwise \emph{t}-tests and then apply a Bonferroni
correction - \strong{none}: computes pairwise \emph{t}-tests and reports the uncorrected statistics}

\item{term}{If \code{NULL}, use each categorical term in the model. Otherwise, only use the given
term.}

\item{alpha}{The family-wise error-rate to restrict the tests to. If "none" is given for
\code{correction}, this value is the value for each test (and is used to calculate the family-wise
error-rate for the group of tests).}

\item{plot}{Setting plot to TRUE will automatically call \code{\link{plot}} on the returned object.}
}
\value{
A list of tables organized by the terms in the model. For each term (categorical terms
only, as splitting on a continuous variable is generally uninformative), the table describes
all of the pairwise-comparisons possible.
}
\description{
Unlike \code{\link{pairwise.t.test}}, this function will return a more informative set of tables similar to
\code{\link{TukeyHSD}}. However, the tables always include the pooled standard error for the estimate and
the degrees of freedom for the test.
}
