% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairwiseReg.R
\name{plotPairwiseReg}
\alias{plotPairwiseReg}
\title{Plot multiple linear regressions for each subgroup for a pair of genes.}
\usage{
plotPairwiseReg(geneA, geneB, datExpr, design_mat)
}
\arguments{
\item{geneA}{The first gene to plot.}

\item{geneB}{The second gene in coexpression with the first gene.}

\item{datExpr}{A gene expression matrix (gene rows by sample columns).}

\item{design_mat}{A model matrix (sample rows by group columns) with 0 or 1 denoting membership in the group.}
}
\value{
A plot of geneA vs geneB expression with subgroup correlations highlighted.
}
\description{
Visualize pairwise gene coexpression correlations for a set of subtypes for two particular genes.
}
