% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tally}
\alias{mean}
\alias{median}
\alias{var}
\alias{sd}
\alias{favstats}
\alias{cohensD}
\alias{cor}
\alias{makeFun}
\alias{xpnorm}
\alias{zscore}
\alias{select}
\alias{filter}
\alias{arrange}
\alias{mutate}
\alias{recode}
\alias{sample}
\alias{resample}
\alias{do}
\alias{shuffle}
\alias{gf_bar}
\alias{gf_boxplot}
\alias{gf_boxploth}
\alias{gf_density}
\alias{gf_histogram}
\alias{gf_histogramh}
\alias{gf_dhistogram}
\alias{gf_dhistogramh}
\alias{gf_point}
\alias{gf_jitter}
\alias{gf_dist}
\alias{gf_lm}
\alias{gf_abline}
\alias{gf_hline}
\alias{gf_vline}
\alias{gf_facet_grid}
\alias{gf_facet_wrap}
\alias{gf_labs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{select}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{recode}}}

  \item{ggformula}{\code{\link[ggformula]{gf_bar}}, \code{\link[ggformula]{gf_boxplot}}, \code{\link[ggformula]{gf_boxploth}}, \code{\link[ggformula]{gf_density}}, \code{\link[ggformula]{gf_histogram}}, \code{\link[ggformula:gf_histogram]{gf_histogramh}}, \code{\link[ggformula:gf_histogram]{gf_dhistogram}}, \code{\link[ggformula:gf_histogram]{gf_dhistogramh}}, \code{\link[ggformula]{gf_point}}, \code{\link[ggformula]{gf_jitter}}, \code{\link[ggformula]{gf_dist}}, \code{\link[ggformula:gf_smooth]{gf_lm}},\code{\link[ggformula:gf_lines]{gf_abline}}, \code{\link[ggformula:gf_lines]{gf_hline}}, \code{\link[ggformula:gf_lines]{gf_vline}}, \code{\link[ggformula:gf_facet_grid]{gf_facet_grid}}, \code{\link[ggformula:gf_facet_grid]{gf_facet_wrap}}, \code{\link[ggformula:gf_aux]{gf_labs}}}

  \item{lsr}{\code{\link[lsr]{cohensD}}}

  \item{mosaic}{\code{\link[mosaic:aggregating]{mean}}, \code{\link[mosaic:aggregating]{median}}, \code{\link[mosaic:aggregating]{var}}, \code{\link[mosaic:aggregating]{sd}}, \code{\link[mosaic:aggregating]{favstats}}, \code{\link[mosaic:aggregating]{cor}}, \code{\link[mosaic]{xpnorm}}, \code{\link[mosaic]{zscore}}, \code{\link[mosaic:resample]{sample}}, \code{\link[mosaic]{resample}}, \code{\link[mosaic]{do}}, \code{\link[mosaic:resample]{shuffle}}}

  \item{mosaicCore}{\code{\link[mosaicCore]{tally}}, \code{\link[mosaicCore]{makeFun}}}
}}
