% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{count}
\alias{tally}
\alias{cor}
\alias{cov}
\alias{favstats}
\alias{IQR}
\alias{mean}
\alias{mean_}
\alias{median}
\alias{max}
\alias{min}
\alias{prod}
\alias{range}
\alias{sd}
\alias{sum}
\alias{var}
\alias{cohensD}
\alias{makeFun}
\alias{xcnorm}
\alias{xpnorm}
\alias{xqnorm}
\alias{xcf}
\alias{xpf}
\alias{xqf}
\alias{zscore}
\alias{select}
\alias{filter}
\alias{arrange}
\alias{mutate}
\alias{recode}
\alias{sample}
\alias{resample}
\alias{do}
\alias{shuffle}
\alias{gf_bar}
\alias{gf_boxplot}
\alias{gf_density}
\alias{gf_histogram}
\alias{gf_dhistogram}
\alias{gf_point}
\alias{gf_jitter}
\alias{gf_dist}
\alias{gf_lm}
\alias{StatLm}
\alias{GeomLm}
\alias{gf_abline}
\alias{gf_hline}
\alias{gf_vline}
\alias{gf_facet_grid}
\alias{gf_facet_wrap}
\alias{gf_labs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{select}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{recode}}}

  \item{ggformula}{\code{\link[ggformula]{gf_bar}}, \code{\link[ggformula]{gf_boxplot}}, \code{\link[ggformula]{gf_density}}, \code{\link[ggformula]{gf_histogram}}, \code{\link[ggformula:gf_histogram]{gf_dhistogram}}, \code{\link[ggformula]{gf_point}}, \code{\link[ggformula]{gf_jitter}}, \code{\link[ggformula]{gf_dist}}, \code{\link[ggformula:gf_smooth]{gf_lm}}, \code{\link[ggformula:ggformula-ggproto]{StatLm}}, \code{\link[ggformula:ggformula-ggproto]{GeomLm}}, \code{\link[ggformula:gf_lines]{gf_abline}}, \code{\link[ggformula:gf_lines]{gf_hline}}, \code{\link[ggformula:gf_lines]{gf_vline}}, \code{\link[ggformula]{gf_facet_grid}}, \code{\link[ggformula:gf_facet_grid]{gf_facet_wrap}}, \code{\link[ggformula:gf_aux]{gf_labs}}}

  \item{lsr}{\code{\link[lsr]{cohensD}}}

  \item{mosaic}{\code{\link[mosaic:aggregating]{cor}}, \code{\link[mosaic:aggregating]{cov}}, \code{\link[mosaic:aggregating]{favstats}}, \code{\link[mosaic:aggregating]{IQR}}, \code{\link[mosaic:aggregating]{mean}}, \code{\link[mosaic:aggregating]{mean_}}, \code{\link[mosaic:aggregating]{median}}, \code{\link[mosaic:aggregating]{max}}, \code{\link[mosaic:aggregating]{min}}, \code{\link[mosaic:aggregating]{prod}}, \code{\link[mosaic:aggregating]{range}}, \code{\link[mosaic:aggregating]{sd}}, \code{\link[mosaic:aggregating]{sum}}, \code{\link[mosaic:aggregating]{var}}, \code{\link[mosaic:xpnorm]{xcnorm}}, \code{\link[mosaic:xpnorm]{xpnorm}}, \code{\link[mosaic:xpnorm]{xqnorm}}, \code{\link[mosaic:cdist]{xcf}}, \code{\link[mosaic:pdist]{xpf}}, \code{\link[mosaic:qdist]{xqf}}, \code{\link[mosaic]{zscore}}, \code{\link[mosaic:resample]{sample}}, \code{\link[mosaic]{resample}}, \code{\link[mosaic]{do}}, \code{\link[mosaic:resample]{shuffle}}}

  \item{mosaicCore}{\code{\link[mosaicCore:prop]{count}}, \code{\link[mosaicCore]{tally}}, \code{\link[mosaicCore]{makeFun}}}
}}

