% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supernova.R
\name{row_term}
\alias{row_term}
\title{Compute and construct an ANOVA table row for a term.}
\usage{
row_term(name, description, models, term)
}
\arguments{
\item{description}{An optional, short description of the term (pedagogical).}

\item{term}{The name of the term the row describes.}

\item{complex}{The complex model containing the term.}

\item{simple}{The simple model (without the term) to compare it to.}
}
\value{
A tibble row with the properties initialized. The code has been written to be as simple
and understanding as possible. Please take a look at the source and offer any suggestions for
improvement!
}
\description{
"Term" is loosely defined here and is probably better understood as "everything in the table that
is not an error row.
}
\keyword{internal}
