% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{select_terms}
\alias{select_terms}
\title{Select terms based on the user's \code{term} specification}
\usage{
select_terms(fit, term = NULL)
}
\arguments{
\item{fit}{A linear model after updating to only include categorical terms.}

\item{term}{The term to check for, if \code{NULL} return all valid terms.}
}
\value{
A character vector of terms to run analyses on.
}
\description{
Before returning the selection, ensure that the term we are subsetting on exists.
}
\keyword{internal}
