\name{Servers}
\docType{data}
\alias{Servers}
\title{Servers data}
\description{
  Data about tips collected from an experiment with 44 servers.

  Note: this dataset will be removed in future versions in favor of \code{\link{Tables}}.
}

\usage{Servers}
\format{
 A dataset with 44 observations on the following 2 variables.
  \describe{
    \item{ServerID}{A number assigned to each server.}
    \item{Tip}{How much the tip was.}
  }
}
