% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{DGP for testing SuperNOVA without mediation}
\usage{
simulate_data(
  n_obs = 1e+05,
  shift_var_index = 1,
  sigma_mod = matrix(c(1, 0.5, 0.8, 0.5, 1, 0.7, 0.8, 0.7, 1), nrow = 3, ncol = 3),
  delta = 1
)
}
\arguments{
\item{n_obs}{Number of observations}

\item{sigma_mod}{Sigma matrix of exposures}

\item{mu}{Vector of means for exposures}

\item{effect_mod_binary}{TRUE/FALSE Whether the effect modifier is binary}

\item{delta_1}{Delta for M1 Exposure}

\item{delta_2}{Delta for M3 Exposure}
}
\description{
DGP for testing SuperNOVA without mediation
}
