% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{DGP for testing SuperNOVA without mediation}
\usage{
simulate_data(
  n_obs = 1e+05,
  shift_var_index = 1,
  sigma_mod = matrix(c(1, 0.5, 0.8, 0.5, 1, 0.7, 0.8, 0.7, 1), nrow = 3, ncol = 3),
  delta = 1
)
}
\arguments{
\item{n_obs}{Number of observations}

\item{shift_var_index}{for the exposure vector that indicates which
exposure to shift}

\item{sigma_mod}{Sigma matrix of exposures}

\item{delta}{Amount to shift exposure by}
}
\value{
A dataframe of simulated data
}
\description{
DGP for testing SuperNOVA without mediation
}
