% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate differentially linearly coexpressed genes between two groups.}
\usage{
simulate_data(samples_per_group = 50, total_genes = 2000,
  maxCor_true = 0.95, minCor_true = 0.3, maxCor_null = 0.2,
  verbose = T, corr_slope = 1, corr_intercept = 0,
  group1_offset = 0, group2_offset = 0)
}
\arguments{
\item{samples_per_group}{The number of samples per group to simulate. Default=50}

\item{total_genes}{The number of genes among all 9 differential coexpression classes to simulate. Will actually return a number of genes which is a multiple of 9 that is less than this number.}

\item{maxCor_true}{The absolute value of the maximum correlation rho value for highly correlated pairs.}

\item{minCor_true}{The absolute value of the minimum correlation rho value for highly correlated pairs.}

\item{maxCor_null}{The absolute value of the maximum correlation rho value for null correlated pairs. The minimum is 0 correlation.}

\item{verbose}{Whether or not to print additional information about the simulation when it runs.}

\item{corr_slope}{The slope of the correlations to simulate. All correlated pairs will have this slope.}

\item{corr_intercept}{The correlation intercept of the genes in matrix B.}

\item{group1_offset}{The correlation intercept of group1 (e.g. its mean expression value)}

\item{group2_offset}{The correlation intercept of group2 (e.g. its mean expression value)}
}
\value{
Returns a list of inputs for chowCor, including a matA, matB, design_mat, subgroup names (conditions), and the actual classes of differential coexpression.
}
\description{
Simulates sets of genes coexpressed with an eigengene which show differential coexpression among two groups. This tool simulates all 9 possible classes of possible gene coexpression changes (positive, negative, null correlation in group A and group B).
}
\keyword{simulate}
\keyword{superNOVA}
