% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mediation_data.R
\name{simulate_mediation_data}
\alias{simulate_mediation_data}
\title{DGP for testing SuperNOVA with mediation}
\usage{
simulate_mediation_data(n_obs = 1e+05, delta = 1)
}
\arguments{
\item{n_obs}{Number of observations}

\item{delta}{Delta for M1 Exposure}

\item{mu}{Vector of means for exposures}
}
\description{
DGP for testing SuperNOVA with mediation
}
