% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{supernova-vctrs}
\alias{supernova-vctrs}
\alias{vec_arith.supernova_number}
\title{Internal \code{vctrs} methods}
\usage{
\method{vec_arith}{supernova_number}(op, x, y, ...)
}
\description{
Internal \code{vctrs} methods
}
\keyword{internal}
