% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperNOVA.R
\name{SuperNOVA}
\alias{SuperNOVA}
\title{Data-adaptive estimation of interactions, effect modification, and
mediation using stochastic shift intervention target parameters. In many mixed exposure settings,
interactions in the mixture, effect modifiers in the covariates that modify the
impact of an exposure and mediating pathways from exposure to outcome are generally unknown. SuperNOVA finds these variable sets
on one part of the data and estimates counterfactual outcome changes given shifts to exposure on an estimation part of the data.
Using cross-validation and targeted learning, estimators are created that utlize machine learning that are unbiased and have the
minimum variance.}
\usage{
SuperNOVA(
  w,
  a,
  z = NULL,
  y,
  deltas,
  estimator = "tmle",
  fluctuation = "standard",
  var_sets = NULL,
  pi_learner = NULL,
  mu_learner = NULL,
  g_learner = NULL,
  e_learner = NULL,
  zeta_learner = NULL,
  n_folds = 2,
  outcome_type = "continuous",
  mediator_type = "continuous",
  quantile_thresh = 0,
  verbose = FALSE,
  parallel = TRUE,
  parallel_type = "multi_session",
  num_cores = 2,
  seed = seed,
  hn_trunc_thresh = 10,
  adaptive_delta = FALSE,
  n_mc_sample = 1000,
  exposure_quantized = TRUE,
  mediator_quantized = FALSE,
  density_type = "sl",
  n_bins = 10,
  max_degree = 2,
  integration_method = "MC",
  use_multinomial = FALSE,
  discover_only = FALSE
)
}
\arguments{
\item{w}{A \code{matrix}, \code{data.frame}, or similar containing a set of
baseline covariates. These variables are measured before exposures.}

\item{a}{\code{matrix}, \code{data.frame}, or similar containing individual or
multiple exposures.}

\item{z}{\code{matrix}, \code{data.frame}, or similar containing individual or
multiple mediators (optional).}

\item{y}{\code{numeric} vector of observed outcomes.}

\item{deltas}{A \code{numeric} value indicating the shift in exposures to
define the target parameter, with respect to the scale of the exposures (A). If adaptive_delta
is true, these values will be reduced.}

\item{estimator}{The type of estimator to fit: \code{"tmle"} for targeted
maximum likelihood estimation, or \code{"onestep"} for a one-step estimator.}

\item{fluctuation}{Method used in the targeting step for TML estimation: "standard" or "weighted".
This determines where to place the auxiliary covariate in the logistic tilting regression.}

\item{var_sets}{A list specifying variable sets for deterministic SuperNOVA usage.
Example: var_sets <- c("A_1", "A_1-Z_2") where the analyst provides variable sets
for exposures, exposure-mediator, or exposure-covariate relationships.}

\item{pi_learner}{Learners for fitting Super Learner ensembles to densities via \pkg{sl3}.}

\item{mu_learner}{Learners for fitting Super Learner ensembles to the outcome model via \pkg{sl3}.}

\item{g_learner}{Learners for fitting Super Learner ensembles to the g-mechanism
g(A|W) (a probability estimator, not a density estimator) for mediation via \pkg{sl3}.}

\item{e_learner}{Learners for fitting Super Learner ensembles to the e-mechanism
g(A|Z,W) (a probability estimator, not a density estimator) for mediation via \pkg{sl3}.}

\item{zeta_learner}{Learners for fitting Super Learner ensembles to the outcome model via \pkg{sl3}..}

\item{n_folds}{Number of folds to use in cross-validation, default is 2.}

\item{outcome_type}{Data type of the outcome, default is "continuous".}

\item{mediator_type}{Data type of the mediator, default is "continuous".}

\item{quantile_thresh}{Threshold based on quantiles of the F-statistic, used to
identify "important" basis functions in the data-adaptive procedure.}

\item{verbose}{Whether to run verbosely (default: FALSE).}

\item{parallel}{Whether to parallelize across cores (default: TRUE).}

\item{parallel_type}{Type of parallelization to use if parallel is TRUE:
"multi_session" (default), "multicore", or "sequential".}

\item{num_cores}{Number of CPU cores to use in parallelization (default: 2).}

\item{seed}{\code{numeric} seed value to be passed to all functions.}

\item{hn_trunc_thresh}{Truncation level for the clever covariate (default: 10).}

\item{adaptive_delta}{If TRUE, reduces the user-specified delta until
the Hn calculated for a shift does not have any observation greater
than hn_trunc_thresh (default: FALSE).}

\item{n_mc_sample}{Number of iterations to be used for the Monte Carlo integration
procedure when using continuous exposures (default: 1000).}

\item{exposure_quantized}{Whether the exposure has been discretized into bins,
in which case the integration procedure is skipped and weighted sums are used instead (default: FALSE).}

\item{mediator_quantized}{If the mediator is discretized, a multinomial ML function
is used in this regression to avoid density estimation (default: FALSE).}

\item{density_type}{Type of density estimation to be used: "sl" for Super Learner
(default) or "hal" for highly adaptive lasso.}

\item{n_bins}{Number of bins for quantizing the exposure if mediation is detected (default: 10).}

\item{max_degree}{Maximum degree of interactions used in the highly adaptive lasso
density estimator if used (default: 1).}

\item{integration_method}{Type of integration to be used in the continuous exposure
case: "MC" for Monte Carlo integration (default) or "AQ" for adaptive quadrature.}

\item{use_multinomial}{Whether to use multinomial regression for binned exposures
(default: FALSE).}

\item{discover_only}{TRUE/FALSE. If TRUE, only the data-adaptive path discovery
is done. No estimates are delivered only exposure mediator sets. If FALSE paths
are both discovered and estimated.}
}
\value{
An S3 object of class \code{SuperNOVA} containing the results of the
procedure to compute a TML or one-step estimate of the counterfactual mean
under a modified treatment policy that shifts a continuous-valued exposure
by a scalar amount \code{delta}. These exposures are data-adaptively
identified using the CV-TMLE procedure.
}
\description{
The SuperNOVA function provides an efficient approach to estimate
interactions, effect modification, and mediation using targeted minimum loss
estimators for counterfactual mean differences under various target parameters.
The procedure employs data-adaptive ensemble b-spline models and stochastic interventions,
leveraging the \pkg{sl3} package for ensemble machine learning. The data is split into V folds, in each fold
the training data is used to find variable sets using flexible basis function estimators. Given the different variable sets,
stochastic intervention target parameters are applied with cross-validated targeted learning.
}
