% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperNOVA.R
\name{SuperNOVA}
\alias{SuperNOVA}
\title{Data-adaptive efficient estimation of interactions and effect
modification using stochastic shift interventions}
\usage{
SuperNOVA(
  w,
  a,
  z = NULL,
  y,
  deltas,
  estimator = "tmle",
  fluctuation = "standard",
  var_sets = NULL,
  pi_learner = NULL,
  mu_learner = NULL,
  g_learner = NULL,
  e_learner = NULL,
  em_learner = NULL,
  zeta_learner = NULL,
  n_folds,
  family = "continuous",
  quantile_thresh = 0.5,
  verbose = FALSE,
  parallel = TRUE,
  parallel_type = "multi_session",
  num_cores = 2,
  seed = seed,
  hn_trunc_thresh = 20,
  adaptive_delta = FALSE
)
}
\arguments{
\item{w}{A \code{matrix}, \code{data.frame}, or similar containing a set of
baseline covariates.}

\item{a}{\code{matrix}, \code{data.frame}, or similar containing an
individual or set of
exposures}

\item{z}{\code{matrix}, \code{data.frame}, or similar containing an
individual or set of mediators}

\item{y}{\code{numeric} vector of the observed outcomes.}

\item{deltas}{A \code{numeric} value indicating the shift in the exposures to
be used in defining the target parameter. This is defined with respect to
the scale of the exposures (A).}

\item{estimator}{The type of estimator to be fit, either \code{"tmle"} for
targeted maximum likelihood or \code{"onestep"} for a one-step estimator.}

\item{fluctuation}{The method to be used in the submodel fluctuation step
(targeting step) to compute the TML estimator. The choices are "standard"
and "weighted" for where to place the auxiliary covariate in the logistic
tilting regression.}

\item{var_sets}{If using SuperNOVA deterministically, this parameter takes
in a list of the form var_sets <- c("A_1", "A_1-Z_2") etc. where the
analyst passes in variable sets for exposures, exosure-mediator, or exposure-
covariate.}

\item{pi_learner}{Learners used to fit Super Learner ensembles to densities
via \pkg{sl3}}

\item{mu_learner}{Learners used to fit Super Learner ensembles to the outcome
model via \pkg{sl3}}

\item{g_learner}{Learners used to fit Super Learner ensembles to the g mech
g(A|W) - a probability estimator not a density estimator used in mediation
via \pkg{sl3}}

\item{e_learner}{Learners used to fit Super Learner ensembles to the e mech
g(A|Z,W) - a probability estimator not a density estimator used in mediation
via \pkg{sl3}}

\item{em_learner}{Super learner from \pkg{sl3} of decision trees used to
apply to the EIF of the shift applied to an exposure using covariates W.}

\item{zeta_learner}{Learners used to fit Super Learner ensembles to the outcome
model via \pkg{sl3}}

\item{n_folds}{Number of folds to use in cross-validation}

\item{family}{Outcome type family}

\item{quantile_thresh}{Threshold based on quantiles of the f-statistic used
to identify "important" basis functions in the data-adaptive procedure}

\item{verbose}{Whether to run verbosely}

\item{parallel}{TRUE/FALSE parallelize across cores}

\item{parallel_type}{If parallel is TRUE, type of parallelization to use. Default
is "multi_session", other values are multicore and sequential.}

\item{num_cores}{Number of CPU cores to use in parallelization}

\item{seed}{\code{numeric} seed value to be passed to all functions}

\item{hn_trunc_thresh}{Truncation level for the clever covariate}

\item{adaptive_delta}{If TRUE, this reduces the user specified delta until
the Hn calculated for a shift does not have any observation that is greater
than hn_trunc_thresh.}
}
\value{
S3 object of class \code{SuperNOVA} containing the results of the
procedure to compute a TML or one-step estimate of the counterfactual mean
under a modified treatment policy that shifts a continuous-valued exposure
by a scalar amount \code{delta}. These exposure are data-adaptively
identified using the CV-TMLE procedure.
)
}
\description{
Data-adaptive efficient estimation of interactions and effect
modification using stochastic shift interventions
}
\details{
Treat variables identified in the same basis functions in ensemble
b-spline models as a data-adaptive parameter. Given variable sets identified,
depending on if basis functions contain variables for \code{A} or \code{AV} construct
targeted minimum loss estimations of the counterfactual mean differences
under various target parameters for individual variables, effect modifying
variables, or interacting variables. Ensemble machine learning may be used
to construct the initial estimates of nuisance functions using \pkg{sl3}.
}
