% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{supernova}
\alias{supernova}
\title{supernova}
\usage{
supernova(fit)
}
\arguments{
\item{fit}{\code{\link{lm}} object}
}
\value{
An object of the class \code{\link{supernova}}, a named list with the munged data.frame and the original \code{\link{lm}} object.
}
\description{
Creates a \code{supernova} object
}
