% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supernova.R
\name{supernova}
\alias{supernova}
\alias{supernova.lm}
\alias{supernova.lmerMod}
\alias{superanova}
\title{supernova}
\usage{
supernova(fit, type = 3, verbose = TRUE)

\method{supernova}{lm}(fit, type = 3, verbose = TRUE)

\method{supernova}{lmerMod}(fit, type = 3, verbose = FALSE)

superanova(fit, type = 3, verbose = TRUE)
}
\arguments{
\item{fit}{A model fit by \code{\link{lm}} or \code{\link[lme4]{lmer}}}

\item{type}{The type of sums of squares to calculate:
\itemize{
  \item \code{1}, \code{I}, and \code{sequential} compute Type I SS.
  \item \code{2}, \code{II}, and \code{hierarchical} compute Type II SS.
  \item \code{3}, \code{III}, and \code{orthogonal} compute Type III SS.
}}

\item{verbose}{If \code{FALSE}, the \code{description} column is suppressed.
Defaults to \code{TRUE}.}
}
\value{
An object of the class \code{supernova}, which has a clean print
  method for displaying the ANOVA table in the console as well as a  named
  list:
  \item{tbl}{The ANOVA table as a \code{\link{data.frame}}}
  \item{fit}{The original \code{\link[stats]{lm}} or \code{\link[lme4]{lmer}}
    object being tested}
  \item{models}{Models created by \code{\link{generate_models}}}
}
\description{
An alternative set of summary statistics for ANOVA. Sums of squares, degrees
of freedom, mean squares, and F value are all computed with Type III sums of
squares. This package adds proportional reduction in error, an explicit
summary of the whole model, and separate formatting of p values and is
intended to match the output used in Judd, McClelland, and Ryan (2017).
}
\details{
\code{superanova()} is an alias of \code{supernova()}
}
\examples{
supernova(lm(Thumb ~ Weight, data = Fingers))

}
\references{
Judd, C. M., McClelland, G. H., & Ryan, C. S. (2017). \emph{Data
  Analysis: A Model Comparison Approach to Regression, ANOVA, and Beyond}
  (3rd ed.). New York: Routledge. ISBN:879-1138819832
}
