% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{supernova}
\alias{supernova}
\alias{superanova}
\title{supernova}
\usage{
supernova(fit)
}
\arguments{
\item{fit}{An \code{\link{lm}} object}
}
\value{
An object of the class \code{\link{supernova}}, which has a clean
print method for displaying the ANOVA table in the console as well as a 
named list:
\item{tbl}{The ANOVA table as a data.frame.}
\item{fit}{The original \code{\link{lm}} object being tested.}
}
\description{
An alternative set of summary statistics for ANOVA. Sums of squares, degrees 
of freedom, mean squares, and F value are all equivalent to 
\code{\link{anova.lm}}. This package adds proportional reduction in error, an
explicit summary of the whole model, and separate formatting of p values and
is intended to match the output used in Judd, McClelland, and Ryan (2017).
}
\details{
`superanova()` is an alias of `supernova()`
}
\examples{
supernova(lm(Thumb ~ Weight, data = Fingers))

}
\references{
Judd, C. M., McClelland, G. H., & Ryan, C. S. (2017). Data 
Analysis: A Model Comparison Approach to Regression, ANOVA, and Beyond 
(3rd edition). New York: Routledge. ISBN:879-1138819832
}
