% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Survey}
\alias{Survey}
\title{Students at a university were asked to enter a random number between 1-20 into a survey.}
\format{
A data frame with 211 observations on the following 1 variable:
\describe{
\item{\code{Any1_20}}{The random number between 1 and 20 that a student thought of.}
}
}
\usage{
Survey
}
\description{
Students at a university taking an introductory statistics course were asked to complete this
survey as part of their homework.
}
\keyword{datasets}
