\name{Tables}
\docType{data}
\alias{Tables}
\title{Tables data}
\description{
  Data about tips collected in an experiment from 44 different tables.
}

\usage{Tables}
\format{
 A dataset with 44 observations on the following 2 variables.
  \describe{
    \item{TableID}{A number assigned to each table.}
    \item{Tip}{How much the tip was.}
  }
}

\examples{
str(Tables)
head(Tables)
}

\keyword{datasets}
