\name{TipExperiment}
\docType{data}
\alias{TipExperiment}
\title{Data from an experiment about smiley faces and tips}
\description{
Data about tips collected in an experiment from 44 different tables.
}
\description{
Servers were randomly assigned to write nothing or "Thank you" on the back of customers' receipts. Data was collected from 44 tables in an effort to examine whether the note would cause more generous tipping.
}

\usage{TipExperiment}
\format{
 A dataset with 44 observations on the following 3 variables.
  \describe{
    \item{TableID}{A number assigned to each table.}
    \item{Tip}{How much the tip was.}
    \item{Condition}{Which experimental condition the table was randomly assigned to.}
  }
}
