\name{TipExperiment}
\docType{data}
\alias{TipExperiment}
\title{Data about tips collected from an experiment with 44 servers.}
\description{
  Servers were randomly assigned to write nothing or "Thank you" on the back of customers' receipts. This experiment was an effort to examine whether the note would cause more generous tipping.
}

\usage{TipExperiment}
\format{
 A dataset with 44 observations on the following 3 variables.
  \describe{
    \item{ServerID}{A number assigned to each server.}
    \item{Tip}{How much the tip was.}
    \item{Condition}{Which experimental condition the server was randomly assigned to.}
  }
}
