% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tables.R
\name{variables}
\alias{variables}
\title{Extract the variables from a model}
\usage{
variables(fit)
}
\arguments{
\item{fit}{An \code{\link{lm}} object of the model with the predictors being tested}
}
\value{
The variables in the model: outcome and predictors
}
\description{
Extract the variables from a model
}
